% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusServer2.R
\name{addToOrder}
\alias{addToOrder}
\title{add condition(s) to the order by clause}
\usage{
addToOrder(order, new, toggle)
}
\arguments{
\item{order}{existing order clause}

\item{new}{new order by field; either "FIELD" or "FIELD desc"}

\item{toggle}{says whether to toggle descending/ascending order}
}
\description{
add condition(s) to the order by clause
}
