% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{api_info}
\alias{api_info}
\title{return information about the api}
\usage{
api_info(env)
}
\value{
a list with these items:
   \itemize{
      \item maxRows; integer maximum number of rows returned by other API calls
   }
}
\description{
return information about the api
}
