% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiveMessage.R
\name{archiveMessage}
\alias{archiveMessage}
\title{Save a handled email message in the appropriate place}
\usage{
archiveMessage(path, valid)
}
\arguments{
\item{path}{character scalar path to file containing the emailwhere the message}

\item{valid}{boolean: is the message valid?
#'}
}
\value{
no return value
'
}
\description{
Both valid and invalid emails are compressed and archived.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

