% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivePath.R
\name{archivePath}
\alias{archivePath}
\title{Save a file or folder in a timestamped location.}
\usage{
archivePath(path, newdir)
}
\arguments{
\item{path}{file or folder to be saved}

\item{newdir}{directory in which to create a timestamped folder which will
hold \code{path} (if it's a file) or the contents of \code{path}, if it's a folder.
The directory \code{newdir} is created, if necessary.

i.e.

  myfile  -> newdir/YYYY-MM-DDTHH-MM-SS/myfile      ## when \code{path} is a file
  mydir/* -> newdir/YYYY-MM-DDTHH-MM-SS_mydir/*     ## when \code{path} is a folder

This means the basename of \code{path} is preserved, either as the
destination filename, or as part of the destination folder name.
This makes it easier to trace the origin of the folder from the
server mainlog.}
}
\value{
TRUE iff the file or folder was successfully archived
}
\description{
Some files such as SG system logs and unhandled files are stored
into folders, possibly for later manual processing.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

