% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{batches_for_all}
\alias{batches_for_all}
\title{get batches for any receiver}
\usage{
batches_for_all(env)
}
\arguments{
\item{batchID}{integer batch ID of largest batch already obtained}

\item{includeTesting}{boolean; default: FALSE.  If TRUE, and the user is an administrator,
then records for batches marked as `testing` are returned as if they were normal batches.}
}
\value{
a data frame with the same schema as the batches table, but JSON-encoded as a list of columns
}
\description{
get batches for any receiver
}
