% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{batches_for_receiver}
\alias{batches_for_receiver}
\title{get batches for a receiver}
\usage{
batches_for_receiver(env)
}
\arguments{
\item{deviceID}{integer device ID}

\item{batchID}{integer batchID; only batches with larger batchID are returned}

\item{includeTesting}{boolean; default: FALSE.  If TRUE, and the user is an administrator,
then records for batches marked as `testing` are returned as if they were normal batches.}
}
\value{
a data frame with the same schema as the batches table, but JSON-encoded as a list of columns
}
\description{
get batches for a receiver
}
