% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{batches_for_receiver_project}
\alias{batches_for_receiver_project}
\title{get batches for a receiver project}
\usage{
batches_for_receiver_project(env)
}
\arguments{
\item{projectID}{integer project ID}

\item{ts}{numeric timestamp}

\item{countOnly}{logical return only the count of records}
}
\value{
a data frame with the same schema as the batches table, but JSON-encoded as a list of columns
}
\description{
get batches for a receiver project
}

