% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltMergeFiles.R
\name{beginsWith}
\alias{beginsWith}
\title{does a string begin with another string}
\usage{
beginsWith(x, y)
}
\arguments{
\item{x}{string in which to look for prefix}

\item{y}{prefix to look for}
}
\description{
does a string begin with another string
}
