% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeSQL.R
\name{$.safeSQL}
\alias{$.safeSQL}
\title{provide read access to some safeSQL internals}
\usage{
\method{$}{safeSQL}(x, name)
}
\arguments{
\item{name}{item name; must be one of 'db' or 'con'}
}
\value{
: either the database filename, or the db connection
}
\description{
provide read access to some safeSQL internals
}

