% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z$_.motusMap.R, R/z[[_.motusMap.R
\name{$<-.motusMap}
\alias{$<-.motusMap}
\title{set an element in a motus map}
\usage{
\method{$}{motusMap}(map, key) <- value

\method{$}{motusMap}(map, key) <- value
}
\arguments{
\item{map}{object of class "motusMap", as created by \code{motusMap}}

\item{key}{key of item}

\item{val}{character string value to associate with \code{key}}

\item{map}{object of class "motusMap", as created by \code{motusMap}}

\item{key}{key of item}

\item{val}{character string value to associate with \code{key}}
}
\description{
Set the value associated with a key in the given map.

Set the value associated with a key in the given map.
}

