% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copse.R
\name{childrenWhere.Copse}
\alias{childrenWhere.Copse}
\title{find children of Twig which satisfy a query}
\usage{
\method{childrenWhere}{Copse}(C, T, expr)
}
\description{
find children of Twig which satisfy a query
}
