% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claimJob.R
\name{claimJob}
\alias{claimJob}
\title{try to accept a job for processing in a queue}
\usage{
claimJob(j, N)
}
\arguments{
\item{j}{the job (an integer of class "Twig")}

\item{N}{integer queue number in the range 1..8}
}
\value{
TRUE if the job has been claimed, FALSE otherwise.
}
\description{
If the given job is still unclaimed, enter it into the specified
queue.  Subsequent attempts to claim the job will fail.  This
operation is atomic; i.e. if there are multiple instances of
processServer(), only one process will succeed in claiming the job.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

