% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTagRegistrations.R
\name{cleanTagRegistrations}
\alias{cleanTagRegistrations}
\title{Fix systematic problems with the tag registration database.}
\usage{
cleanTagRegistrations(m, s, cleanBI = FALSE)
}
\arguments{
\item{m}{tbl with full tag info from motus, as returned by
motusSearchTags()}

\item{s}{sqlite_src database where tables "tags" and "events" will
be created.}

\item{cleanBI}{logical; if TRUE, attempt to clean the burst interval;
otherwise, leave it as-is.  Default: FALSE.}
}
\value{
returns modified m; side effect: creates or overwrite the table "tags"
in \code{s}, which will contain cleaned tag registrations.  Also
creates or overwrites table "events", giving activation / deactivation
timestamps for each tag deployment.

The \code{events} table looks like this:
\strong{events:}
\itemize{
\item ts    timestamp for event
\item tagID motus tag ID for event
\item event integer: 1 is activation; 0 is deactivation
}

path to an sqlite database usable by the tag finder; it will have these tables:

\strong{tags:}
\itemize{
\item tagID      motus tag ID
\item nomFreq    nominal frequency, in MHz e.g. 166.38
\item offsetFreq offset from nominal, in kHz
\item param1     first interpulse gap
\item param2     second interpulse gap
\item param3     third interpulse gap
\item period     burst interval
\item mfgID      lotek ID code
\item codeSet    Lotek codeset name
}
}
\description{
Some tag registrations had problems.  This All available motus metadata are
The list of all registered tags is obtained from the motus-wts.org
server.  A cleaned-up database suitable for the find_tags_motus
program is generated, including an events table which indicates
when tags were activated and inactivated.  This allows the database
to be used against any receiver dataset, regardless of the dates.
}
\note{
:

For Lotek coded ID tags, the registrations are cleaned up like so:
\itemize{

\item empirical gap values are replaced with nominal values from
the Lotek codeset

\item burst intervals are replaced by the mean of nearby good
  values.  Nearby means within 0.05s, and good means no more than
  0.0005 s shorter than the longest nearby BI.

}

Registration problems appear to be mainly from dropped USB packets,
when the computer used to make tag recordings has not been able to
keep up with the full funcubedongle sampling rate.  The result is
overly variable estimates of gap values and burst intervals, with a
bias downward from the true values (since dropped packets represent
lost time).

The tag activation events are generated using these items from the
motus database, in order of preference (i.e. the first available
item is used):

\enumerate{

\item tsStart - the starting date for a tag deployment record;
tsStartCode = 1L

\item dateBin - the start of the quarter year in which the tag was
expected to be deployed;
tsStartCode = 2L

\item ts - the date the tag was registered;
tsStartCode = 3L

}

Tag deactivation events are generated using these items, again
in order of preference:

\enumerate{

\item tsEnd - the ending date for a tag deployment; e.g. if a tag
was found, or manually deactivated; tsEndCode = 1L

\item tsStart for a different deployment of the same tag; tsEndCode = 2L

\item tsStart + predictTagLifespan(model, BI) * marginOfError
if the tag model is known; tsEndCode = 3L

\item tsStart + predictTagLifespan(guessTagModel(speciesID), BI) * marginOfError
if the species is known; tsEndCode = 4L

\item 90 days if no other information is available; tsEndCode = 5L

}

Also, as of 6 April 2016, we're using a lifetime of 700 days for tags in the
Taylr 2013 project (gulls)
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
