% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeRecvSrc.R
\name{closeRecvSrc}
\alias{closeRecvSrc}
\title{Close the src_sqlite for a receiver database.}
\usage{
closeRecvSrc(src)
}
\arguments{
\item{src}{dplyr::src_sqlite as returned by \link{\code{getRecvSrc()}}}
}
\value{
no return value.
}
\description{
receiver database files are stored in a single directory, and
have names like "SG-XXXXBBBKYYYY.motus" or "Lotek-NNNNN.motus"
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
