% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commitMetadataHistory.R
\name{commitMetadataHistory}
\alias{commitMetadataHistory}
\title{commit any changes to the metadata history git repo}
\usage{
commitMetadataHistory(meta)
}
\arguments{
\item{meta:}{safeSQL object to DB containing table \code{meta} to
which the metadata commit hash will be written as a key-value
pair \code{c('hash', hash)}.}
}
\value{
TRUE on success.  Generates an error on failure.
}
\description{
commit any changes to the metadata history git repo
}
\details{
Any changes to the files in the metadata history git repository
are committed and pushed upstream.
}
\note{
This function uses \code{\link{lockSsymbol('metadata_history')}} to protect
access to the git repo, but this function \emph{must} be called within an EXCLUSIVE transaction
on \code{meta} that is both begun and committed by the caller.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
