% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareOldNew.R
\name{compareOldNew}
\alias{compareOldNew}
\title{compare new-style tag-finder results with the traditional approach}
\usage{
compareOldNew(year, proj, site, oldSym = 25, newSym = 24)
}
\arguments{
\item{year}{- integer year}

\item{proj}{- sensorgnome project code}

\item{site}{- sensorgnome site code}

\item{oldSym}{- plot symbol for old detections; default:  25 (downward triangle)}

\item{newSym}{- plot symbol for new detections; default: 24 (upward triangle)}
}
\value{
a character vector of full paths to any files created (plots, datasets);
a vector of length zero if no plots or datasets are produced.
}
\description{
Plot a Year/Project/Site tag detection plot (condensed to first detection
of each tag per hour), overplotting detections by old and new methods.
}
\note{
: For projects marked "PRIVATE", this function uses both private and public
tag detections, so results should be treated appropriately.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

