% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgSyncDBtoRepo.R
\name{compressedFileDone}
\alias{compressedFileDone}
\title{is a .gz file actually finished?}
\usage{
compressedFileDone(gzfile)
}
\arguments{
\item{gzfile}{character scalar filename}
}
\value{
TRUE if the file exists and is a complete (valid) .gz file;
FALSE otherwise
}
\description{
is a .gz file actually finished?
}

