% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFilterTagsDB.R
\name{createFilterTagsDB}
\alias{createFilterTagsDB}
\title{create a tag database for use by the dta2sg function in the sensorgnome
package}
\usage{
createFilterTagsDB(tsRange, outFile)
}
\arguments{
\item{outFile}{name of output file for database}

\item{ts}{real vector of length 2; start and end timestamps;  any tag
active during the time period is included in the output database.}
}
\value{
the number of tags in the database
}
\description{
(Deprecated)
}
\details{
End users who wish to filter .DTA files themselves can use the database
generated by this function.  It will include all tags whose activation overlaps
the specified time interval.
}
\note{
the output database is a .csv file with these columns:
\itemize{
\item proj: character name of project
\item id: Lotek id tag ID, but possibly with a fractional part to distinguish
tags with different BI from the same project; e.g. 123.1
\item tagFreq: floating point nominal tag frequency, in MHz
\item bi: floating point burst interval, in seconds
}
sorted in order by id and bi within project.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
