% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{dataServer}
\alias{dataServer}
\title{serve http requests for tag detection data}
\usage{
dataServer(port = 55930, tracing = FALSE)
}
\arguments{
\item{port}{integer; local port on which to listen for requests
Default: 0xda7a}

\item{tracing}{logical; if TRUE, run interactively, allowing local user
to enter commands.}
}
\value{
does not return; meant to be run as a server.
}
\description{
serve http requests for tag detection data
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
