% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransferRegex.R
\docType{data}
\name{dataTransferRegex}
\alias{dataTransferRegex}
\title{A PCRE-style regular expression for matching URLs in an email
message sent to share data.}
\format{An object of class \code{character} of length 1.}
\usage{
dataTransferRegex
}
\description{
A PCRE-style regular expression for matching URLs in an email
message sent to share data.
}
\keyword{datasets}

