% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteFromMotus.R
\name{deleteFromMotus}
\alias{deleteFromMotus}
\title{delete a set of detections from motus by adding records
to the batchDelete table.}
\usage{
deleteFromMotus(src, reason = "not given", monoBN = NULL, tsStart = 0,
  tsEnd = 1e+11)
}
\arguments{
\item{src}{dplyr src_sqlite to receiver database}

\item{reason}{character scalar giving reason for deletion, in human-readable form}

\item{monoBN}{boot session number; required if an SG receiver; omit
or set to NULL for a Lotek receiver.}

\item{tsStart}{starting timestamp; if specified, only those batches
containing at least 1 detection with timestamp >= tsStart are deleted}

\item{tsEnd}{ending timestamp; if specified, only those batches
containing at least 1 detection with timestamps <= tsEnd are deleted}
}
\value{
TRUE on success; FALSE otherwise
}
\description{
delete a set of detections from motus by adding records
to the batchDelete table.
}
\details{
detections are added or removed from motus in batches.  For
Lotek receivers, a batch is a sequence of all detections in a range
of timestamps.  For an SG receiver, a batch is additionally constrained
to be within a single boot session.  Boot sessions can cover multiple
batches, however.

The batches deleted will be:
\enumerate{
\item Lotek receiver: all batches satisfying the tsStart and/or tsEnd
constraints, whichever are specified, are deleted.  If neither tsStart
nor tsEnd is specified, all batches for this receiver are deleted.
\item SG receiver: all batches with the specified monoBN and satsifying
any specified tsStart / tsEnd constraints are deleted.  If neither
tsStart nor tsEnd is specified, all batches for this receiver and
boot session are deleted.
}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

