% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteTestBatches.R
\name{deleteTestBatches}
\alias{deleteTestBatches}
\title{delete one or more test batches from the master motus DB}
\usage{
deleteTestBatches(batchID)
}
\arguments{
\item{batchID}{integer vector of IDs of batches to be deleted; only those
which are marked in the DB as testing (i.e. whose `status` field is -1)
will be deleted.}
}
\value{
a logical vector of the same length as `batchID`, indicating which
batches were successfully deleted.
}
\description{
delete one or more test batches from the master motus DB
}
\details{
test batches are generated when a top-level job includes a
TRUE `isTesting` parameter.  This is used for debugging.

Once such batches are in the master database, they will only be
returned by the dataServer for API `batches_for_*` calls that have
`includeTesting` set to TRUE, and only for admin users.

Regardless, it will sometimes be useful to remove test batches from
the master DB.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
