% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digestFile.R
\name{digestFile}
\alias{digestFile}
\title{calculate a digest of a file}
\usage{
digestFile(f, algo = "sha1")
}
\arguments{
\item{f}{character scalar; path to file}

\item{algo;}{algorithm for digest; default: "sha1"}
}
\value{
character scalar lower-case hex sha1 digest
}
\description{
calculate a digest of a file
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
