% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirSortedBy.R
\name{dirSortedBy}
\alias{dirSortedBy}
\title{Get listing of full paths to items in folder(s), sorted by choice of
item from file.info}
\usage{
dirSortedBy(paths, sortBy = "ctime", recursive = FALSE)
}
\arguments{
\item{paths}{set of folders to list}

\item{sortBy}{character vector; names of one or more columns
    returned by file.info:

\itemize{
   \item size
   \item isdir
   \item mode
   \item mtime
   \item ctime
   \item atime
   \item uid
   \item gid
   \item uname
   \item grname
}
or these additional names:
\itemize{
   \item name - filename, without directory
   \item path - full path to file
}}

\item{recursive}{boolean; should all items from subfolders also be listed?}
}
\value{
a character vector of paths to files, sorted by the specified property.
The sorting happens *across* folders, rather than within them.
}
\description{
Get listing of full paths to items in folder(s), sorted by choice of
item from file.info
}
\seealso{
\link{\code{dir}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
