% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.dropbox.R
\name{download.dropbox}
\alias{download.dropbox}
\title{Grab a file or folder specified by a dropbox shared link.}
\usage{
download.dropbox(link, dir)
}
\arguments{
\item{link}{URL of file or folder on dropbox.com, from the email
sent to the file or folder recipient.}

\item{dir}{directory into which the file(s) will be downloaded}
}
\value{
returns invisible(NULL)
}
\description{
Download the file or folder (recursively) from a URL pointing to
shared content on dropbox.com into the specified directory.
}
\note{
This will not work with the URL in the email that dropbox
    generates when a user chooses to share a file in the most
    obvious (to me) way!  Rather, the user must explicitly create a
    link that permits anyone to view the file, then manually email
    it to us.  We have filed a feature request on this [1].
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\references{
\link{https://www.dropboxforum.com/hc/en-us/community/posts/206649516-Import-Email-shared-file-by-api-?page=1#community_comment_212400843}
}

