% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.FTP.R
\name{download.FTP}
\alias{download.FTP}
\title{Download files from an FTP site.}
\usage{
download.FTP(link, dir, suffixes = c(".gz", ".txt", ".zip", ".7z", ".rar",
  ".ZIP", ".7Z", ".TXT", ".RAR", ".GZ"), maxDepth = 5)
}
\arguments{
\item{link}{URL of an FTP location.  If the site is password protected, the URL
must contain a \code{USER:PASSWORD@} portion (see the example)}

\item{dir}{directory into which the file(s) will be downloaded}

\item{suffixes}{vector of file types (by suffix) to download.
Defaults to \code{c(".gz", ".txt", ".zip", ".7z")}.}

\item{maxDepth}{maximum depth of nested folders to download, starting
at \code{URL}.  Default: 5.}
}
\value{
returns invisible(NULL)
}
\description{
Recursively downloads all files from the specified URL.
The permitted file suffixes and maximum recursion depth can be set to protect
against accidentally downloading far more than was intended.
}
\examples{

## download.FTP("ftp://USER:PASSWD@ftp.depot.qc.ec.gc.ca/depot/SG/Mai2016/BBBK0489_Estimauville")

}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

