% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.wetransferConf.R
\name{download.wetransferConf}
\alias{download.wetransferConf}
\title{Download a file specified by a wetransfer.com "confirmation email" link.}
\usage{
download.wetransferConf(link, dir)
}
\arguments{
\item{link}{URL of file on wetransfer.com, from confirmation email
sent to file sender.}

\item{dir}{directory into which the file(s) will be downloaded}
}
\value{
a messages saying how many bytes were downloaded.
}
\description{
The confirmation email sent from wetransfer.com to a file sender
contains a shortened link to the downloadable file.  That file is
downloaded into the specified directory.
}
\note{
wetransfer.com does not have a published download API, so we
    do this the tedious way, by parsing responses from their
    server.  Watch for changes to the format of emails and server
    replies that might break this fragile code.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{download.wetransferDirect}} for downloading using
the link in an email to the file recipient.
}

