% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.wetransferDirect.R
\name{download.wetransferDirect}
\alias{download.wetransferDirect}
\title{Download a file specified by a direct wetransfer.com link.}
\usage{
download.wetransferDirect(link, dir)
}
\arguments{
\item{link}{URL of file on wetransfer.com, from email sent to the file
recipient.}

\item{dir}{directory into which the file(s) will be downloaded}
}
\value{
returns invisible(NULL)
}
\description{
The email sent from wetransfer.com to a file recipient contains a
direct link to the downloadable file.  That file is downloaded
into the specified directory.
}
\note{
wetransfer.com does not have a published download API, so we
    do this the tedious way, by parsing responses from their
    server.  Watch for changes to the format of emails and server
    replies that might break this fragile code.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{download.wetransferConf}} for downloading using
the link in a confirmation email to the file sender.
}

