% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGoogleDrive.R
\name{downloadGoogleDrive}
\alias{downloadGoogleDrive}
\title{Download a file or folder specified by a google drive shared link.}
\usage{
downloadGoogleDrive(link, dir)
}
\arguments{
\item{link}{URL of file on drive.google.com}

\item{dir}{directory into which the file(s) will be downloaded}
}
\value{
stdout and stderr from running \code{gdrive}
}
\description{
An email from a user via Google Drive offering a file or folder for
sharing includes a URL to that item and requires no further
credentials.  This URL is downloaded recursively into a directory
using the gdrive command-line client (see references).
}
\references{
\url{https://github.com/prasmussen/gdrive}

\url{https://developers.google.com/drive/v3/web/about-sdk}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
