% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadWetransferDirect.R
\name{downloadWetransferDirect}
\alias{downloadWetransferDirect}
\title{Download a file specified by a direct wetransfer.com link.}
\usage{
downloadWetransferDirect(link, dir)
}
\arguments{
\item{link}{URL of file on wetransfer.com, from email sent to the file
recipient.}

\item{dir}{directory into which the file(s) will be downloaded}
}
\value{
character scalar saying how many bytes were downloaded.  It has
these attributes:
\itemize{
\item \code{fileName} name of the downloaded file
\item \code{hashName} security-hash version of the downloaded file's
name, e.g. "wetransfer-ab2de9.zip"  When multiple files are sent in
a single transfer, \code{fileName==hashName}.  The hashName is
convenient for recording which downloads have been made.
}
}
\description{
The email sent from wetransfer.com to a file recipient contains a
direct link to the downloadable file.  That file is downloaded
into the specified directory.
}
\note{
wetransfer.com does not have a published download API, so we
    do this the tedious way, by parsing responses from their
    server.  Watch for changes to the format of emails and server
    replies that might break this fragile code.
}
\seealso{
\link{\code{download.wetransferConf}} for downloading using
the link in a confirmation email to the file sender.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
