% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusServer.R
\name{dumpJobDetails}
\alias{dumpJobDetails}
\title{dump details of job j to res, as ith job listing}
\usage{
dumpJobDetails(res, j, i)
}
\arguments{
\item{res}{Rook::response object}

\item{j}{job}

\item{i}{integer; index in list of jobs to be displayed}
}
\description{
dump details of job j to res, as ith job listing
}
