% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{email}
\alias{email}
\title{Send an email message.}
\usage{
email(to, subj, msg, ..., attachment = NULL)
}
\arguments{
\item{to}{email address of recipient}

\item{subj}{character scalar subject line}

\item{msg}{character scalar message content. If further parameters
are specified in \code{...}, this is treated as a
\code{sprintf}-style formatting string, with fields filled in
from \code{...}.  Otherwise, it is used as-is.}

\item{...}{parameters for replacing \code{sprintf} formatting codes
in \code{msg}}

\item{attachment}{named list of attachments; the list items are paths
to files to attach, and the list names are the corresponding labels
as they will be seen by the recipient.  Default: NULL}
}
\value{
invisible(NULL)
}
\description{
An email message is sent from data@sensorgnome to the specified user.
A log entry is made, and the outgoing message is saved in the motus
server outbox.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{sprintf}} for formatting codes.
}

