% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embroilHuman.R
\name{embroilHuman}
\alias{embroilHuman}
\title{Save a set of files we don't know how to deal with, and request human
intervention.}
\usage{
embroilHuman(files, msg = "")
}
\arguments{
\item{files}{character vector of full paths to files}

\item{msg}{additional character vector to include in the email message}
}
\value{
invisible(NULL)
}
\description{
Save a set of files we don't know how to deal with, and request human
intervention.
}
\details{
a full directory structure with any files of unknown type
    from an email, download, or manual folder copy is saved into a
    subfolder of \code{MOTUS_PATH$MANUAL}.  If all files are in the
    same directory, that directory is simply moved into
    \code{MOTUS_PATH$MANUAL}.  Otherwise, the greatest common
    parent path is removed from all file paths, and the files are
    moved to a new folder, retaining the non-shared portion of
    their paths.

An email is sent to the poor chump whose job it is to deal with
this.
}
\seealso{
\link{\code{server}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
