% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enqueueCommand.R
\name{enqueueCommand}
\alias{enqueueCommand}
\title{Enqueue a command that doesn't require a file or folder.}
\usage{
enqueueCommand(name, ...)
}
\arguments{
\item{name}{command name; e.g. 'SGnew'}

\item{...}{additional parameters for the command}
}
\value{
TRUE on success.
}
\description{
Enqueue a command that doesn't require a file or folder.
}
\details{
An empty file is created with the name \code{paste0(TIMESTAMP, MOTUS_QUEUE_SEP, paste0(c(name, ...), collapse=MOTUS_QUEUE_SEP))}.  This file is moved to the processing queue.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

