% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureMotusTransferTables.R
\name{ensureMotusTransferTables}
\alias{ensureMotusTransferTables}
\title{Ensure we have the mysql transfer tables for moving SG data to
motus.}
\usage{
ensureMotusTransferTables(src, recreate = c())
}
\arguments{
\item{src}{dplyr::src_mysql to motus transfer database, e.g. as returned
by \link{\code{openMotusDB}}}

\item{recreate}{vector of table names which should be dropped then re-created,
losing any existing data.  Defaults to empty vector, meaning no tables
are recreate.}
}
\description{
Ensure we have the mysql transfer tables for moving SG data to
motus.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

