% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureRecvDBTables.R
\name{ensureRecvDBTables}
\alias{ensureRecvDBTables}
\title{make sure a receiver database has the required tables; also, load
custom SQLite extensions on this DB connection.}
\usage{
ensureRecvDBTables(src, recreate = c(), serno)
}
\arguments{
\item{src}{dplyr sqlite src, as returned by \code{dplyr::src_sqlite()}}

\item{recreate}{vector of table names which should be dropped then re-created,
losing any existing data.  Defaults to empty vector, meaning no tables
are recreate.  As a special case, TRUE causes all tables to be dropped
then recreated.}

\item{serno}{character scalar receiver serial number; only used when `meta`
table is populated}
}
\value{
returns NULL (silently); fails on any error
}
\description{
make sure a receiver database has the required tables; also, load
custom SQLite extensions on this DB connection.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
