% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureServerDB.R
\name{ensureServerDB}
\alias{ensureServerDB}
\title{ensure we have created the server database}
\usage{
ensureServerDB()
}
\value{
safeSQL connection to the server database.
}
\description{
This database holds a table of symbolic locks (e.g. to prevent multiple
processes from running a job on the same receiver in parallel)
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

