% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureServerTables.R
\name{ensureServerTables}
\alias{ensureServerTables}
\title{ensure we have database tables for running the server.}
\usage{
ensureServerTables(recreate = c())
}
\arguments{
\item{recreate}{vector of table names which should be dropped then re-created,
losing any existing data.  Defaults to empty vector, meaning no tables
are recreate.  As a special case, TRUE causes all tables to be dropped
then recreated.}
}
\value{
returns NULL (silently); fails on any error
}
\description{
ensure we have database tables for running the server.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

