% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_from_app.R
\name{error_from_app}
\alias{error_from_app}
\title{return an error from a Rook app}
\usage{
error_from_app(error, ...)
}
\arguments{
\item{error}{character scalar error message to return to app caller}

\item{...}{additional parameters describing the error.  These are
JSON-encoded into the API response.}
}
\value{
Similar to the return value from
    \link{\code{return_from_app}}, except blessed with class
    "error", to make it easy for rook apps to check for an error;
    e.g. \code{if (inherits(X, "error")) return(X)}
}
\description{
This function is called to generate an in-band error reply to the API
request.  The remote requestor will still receive a bzip2-compressed
JSON-formatted object, but that object will have a single field
named "error", whose value is the string passed to this function.
}
\details{
Objects returned from successful API calls \emph{never} include a
top-level field called "error".
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
