% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportGlobalTags.R
\name{exportGlobalTags}
\alias{exportGlobalTags}
\title{export a global tags view as an old-style globaltags .rds file}
\usage{
exportGlobalTags(projectID, create = TRUE,
  exportFolder = MOTUS_PATH$TAG_PROJ)
}
\arguments{
\item{projectID}{integer scalar; motus project ID}

\item{create}{if TRUE, recreate the tagProj .sqlite database from scratch;
otherwise, use the existing one.}

\item{exportFolder}{character scalar; default:  \code{MOTUS_PATH$TAG_PROJ}}
}
\value{
no return value; writes a file called "proj_NNN_globaltags.rds" to
\code{exportFolder}
}
\description{
This function is a stop-gap to allow users to get their full data sets
in advance of the full end-user motus R package being available.
}
\seealso{
\code{\link{getGlobalTagsView}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
