% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusServer2.R
\name{files_from_recv_DB}
\alias{files_from_recv_DB}
\title{return a dplyr::tbl of files from a receiver database}
\usage{
files_from_recv_DB(serno, day = NULL)
}
\arguments{
\item{serno}{character scalar receiver serial number}

\item{day}{character scalar day; default NULL}
}
\value{
a tbl; one of two flavours, depending on
the receiver type and `day`.

If the receiver is a sensorgnome and day is \code{NULL}, the return value has
these columns sorted by decreasing day:
\itemize{
\item day: character; day, formatted as "YYYY-MM-DD"
\item count: integer; number of files from the day in the receiver DB `files` table.
}
Otherwise, for Lotek receivers or if `day` is given, the return value has these columns,
sorted by fileID:
\itemize{
  \item fileID: integer; ID of file
  \item name: character; name of file
  \item bootnum: integer; boot count, uncorrected; if an SG file
  \item monoBN: integer; corrected boot count; if an SG file
  \item contentSize: integer; uncompressed file size in bytes
  \item jobID: the integer motus ID for the job in which this file was most recently updated
}
}
\description{
return a dplyr::tbl of files from a receiver database
}
