% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusServer2.R
\name{files_from_repo}
\alias{files_from_repo}
\title{return a dplyr::tbl of files from the file repository
for a receiver (and maybe specific day)}
\usage{
files_from_repo(serno, day = NULL)
}
\arguments{
\item{serno}{character scalar receiver serial number}

\item{day}{character scalar day; default NULL}
}
\value{
a tbl; one of two flavours, depending on
the receiver type and `day`.

If the receiver is a sensorgnome and day is \code{NULL}, the return value has
these columns sorted by decreasing day:
\itemize{
\item day: character; day, formatted as "YYYY-MM-DD"; any days for which the repo has files
\item count: integer; number of files from the day in the file repo
}

Otherwise, for Lotek receivers or if `day` is given, the return value has these columns,
sorted by fileID:
\itemize{
  \item name: character; name of file
  \item fileSize: integer; size of file on disk, in bytes
}
For Lotek receivers, day is ignored (as long as it is present), and all files for the
receiver are listed.
}
\description{
return a dplyr::tbl of files from the file repository
for a receiver (and maybe specific day)
}
\note{
for SG receivers, when both a compressed and un uncompressed version of the file
are present, this only counts as one file, and if returning files, the name is returned
without a ".gz" suffix, but the size is of the .gz file on disk
}
