% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixDOSfilenames.R
\name{fixDOSfilenames}
\alias{fixDOSfilenames}
\title{Correct DOS-style 8.3 filenames into their original long filenames.
Any files needing renaming are renamed on disk.}
\usage{
fixDOSfilenames(f, info)
}
\arguments{
\item{f}{vector of full paths to files}

\item{info}{data frame of split filename components, which are the
named capture groups in \link{\code{sgFilenameRegex}}}
}
\value{
info
}
\description{
Some users move SG data files along a route that fails to preserve
their long filenames, forcing them into a DOS-style 8.3 filename.
Typically, this only happens to a few files per batch, for some reason.
Perhaps they are storing raw files in the top level folder of a VFAT
filesystem, which limits the total length of filenames in a folder.
}
\details{
In any case, these can be recognized by a tilde \code{~} in the
filename, and the first two letters of the shortened name will
match (case insensitively) the first two letters of the original
name.
}
\note{
Corrections are performed as so:

\itemize{

\item if there are files with unaltered long names from exactly one
sensorgnome in \code{f}, then the shortened files are assumed to
have come from that receiver, and their names are corrected
post-hoc using content timestamps.

\item otherwise, take the the first two characters of each
shortened name and see whether they match (case insensitively) the
first two characters of unshortened names of a single receiver.
Any shortened names for which that is true are corrected.

\item otherwise, for any remaining shortened filenames, we can't
tell which receiver(s) the misnamed files belong to, so they are
saved in a subfolder of "/sgm/manual" with a "README.TXT" giving
details.  FIXME: get first and last timestamps in shortened files
and compare to timestamps parsed from full names.

}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

