% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusServer2.R
\name{get_job_stackdump}
\alias{get_job_stackdump}
\title{return a URL to an .rds file stack dump for a job with errors}
\usage{
get_job_stackdump(env)
}
\description{
return a URL to an .rds file stack dump for a job with errors
}
