% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDownloadURL.R
\name{getDownloadURL}
\alias{getDownloadURL}
\title{return the URL to the download site for a project}
\usage{
getDownloadURL(projectID, errorJobID, isTesting = FALSE)
}
\arguments{
\item{projectID}{[optional] integer scalar motus project ID}

\item{errorJobID}{[optional] integer scalar job ID}

\item{isTesting}{logical scalar; if TRUE, the URL is to
the testing hierarchy of web products, rather than
the usual one.  Default:  FALSE}
}
\value{
character scalar giving URL to:
\itemize{
\item download site for project, if `projectID` is specified
\item or .rds stack dump file for job with errors, if `errorJobID` is specified
}
}
\description{
return the URL to the download site for a project
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
