% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeeder.R
\name{getFeeder}
\alias{getFeeder}
\title{return a function which returns the next item from a folder.}
\usage{
getFeeder(incoming, tracing = FALSE)
}
\arguments{
\item{incoming}{the full path to the incoming folder.}

\item{tracing}{boolean; if TRUE, each event on the incoming folder
is printed.  Default: FALSE.}
}
\value{
a function with no parameters.  This function returns the
    full path to the next available incoming item, or waits if
    there are none.
}
\description{
The returned function maintains a queue of path names which is
initialized with those files and folders already in
\code{incoming}.  Any files or folders subsequently added to
\code{incoming} are added to the end of the queue.  Items are
typically email messages with attached data files or links to
download them, or folders of data files already on the server.
}
\details{
The algorithm is this:

\enumerate{
\item initialize file queue with list of files and folders in \code{incoming}.
\item use inotifywait to watch \code{incoming} for new files, links, and
folders.
}

There's a race condition in which files might be created after
starting to watch \code{incoming} but before calling \code{dir()}
to list existing items, so that we see them twice.  Deal with this
by keeping track of existing files.  i.e. the queue looks like
this:

1     ---+
2        |
...      +--- files existing before inotifywait is active
n     ---+

n + 1 ---+
n + 2    |
n + 3    +--- files created after inotifywait is active but before dir()
...      |    These files are seen twice:  once from calling dir()...
n + m ---+

n + m + 1 --+
n + m + 2   |
...         +--- ... and this second time, from the inotifywait pipe
n + 2m   ---+

n + 2m + 1 ---+
n + 2m + 2    |--- files created after dir()
...        ---+

m will usually be zero, but that can't be guaranteed.  We retain the list
of items 1..n + m, and check inotify events against it. The first item
not in that list will be item n + 2m + 1, which is new.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{server}}
}

