% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileAsBZ2.R
\name{getFileAsBZ2}
\alias{getFileAsBZ2}
\title{read a file and (re-) compress it into a bzip2-format raw vector}
\usage{
getFileAsBZ2(f, ext = stri_extract_last_regex(f, "\\\\.[A-Za-z0-9]*$"),
  size = file.info(f)$size)
}
\arguments{
\item{f}{full path to file.  This can be an uncompressed file, or
one compressed with gzip (extension must be ".gz") or bzip2 (extension
must be ".bz2").}

\item{ext}{[optional] file extension.  If not supplied, the
extension is the substring of \code{f} beginning with the last period (\code{'.'}),
for example ".gz".  If NA, the file is treated as not compressed.}

\item{size}{[optional] size of file on disk. If not supplied, it
is obtained by a call to file.info}
}
\value{
a raw vector of bzip2-compressed data, with attribute \code{"len"},
giving the length of the uncompressed data.
}
\description{
read a file and (re-) compress it into a bzip2-format raw vector
}
\note{
If the file \code{f} is empty, this returns a raw vector of length 0.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
