% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileAsText.R
\name{getFileAsText}
\alias{getFileAsText}
\title{read a file as text, converting if necessary}
\usage{
getFileAsText(f)
}
\arguments{
\item{f}{full path to file.}
}
\value{
a vector of lines in the file, after converting to text format
}
\description{
read a file as text, converting if necessary
}
\details{
The type of the file is determined using the linux \code{file}
utility.  These type conversions are known:

\itemize{
\item Rich Text Format: the file is converted using \code{unrtf --text},
and the leading comments are stripped.
\item ASCII text: the file is read as-is.
}
}
\note{
If the file \code{f} is empty, this returns a vector of length 0.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

