% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadataHash.R
\name{getMetadataHash}
\alias{getMetadataHash}
\title{get the commit hash for the currently-cached motus metadata}
\usage{
getMetadataHash()
}
\value{
character scalar commit hash, a hex string with 40 characters.
}
\description{
get the commit hash for the currently-cached motus metadata
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
