% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotusDeviceID.R
\name{getMotusDeviceID}
\alias{getMotusDeviceID}
\title{Get the motus device ID for a receiver, given its database.}
\usage{
getMotusDeviceID(src, useFirst = FALSE)
}
\arguments{
\item{src}{dplyr src_sqlite to receiver database, or a character
serial number}

\item{useFirst}{if TRUE, the first receiver matching the given
serial number is used. if FALSE and more than one
receiver matches, fail with an error.}
}
\value{
the sensor ID, an integer
}
\description{
Get the motus device ID for a receiver, given its database.
}
\note{
As a side-effect, the motus device ID is also stored in the
receiver database "meta" table, with key "deviceID"
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
