% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotusMetaDB.R
\name{getMotusMetaDB}
\alias{getMotusMetaDB}
\title{Get the motus database of metadata for tags, receivers, projects,
and species.}
\usage{
getMotusMetaDB()
}
\value{
a path to an sqlite database usable by the \code{tagview} function
It will have these tables:

\strong{tags:}
\itemize{
\item tagID motus tag ID
\item projectID
\item mfgID
\item dateBin
\item type
\item codeSet Manufacturer codeset name
\item manufacturer
\item model
\item lifeSpan in days
\item nomFreq  nominal frequency, in MHz e.g. 166.38
\item offsetFreq offset from nominal, in kHz
\item period  burst interval
\item periodSD
\item pulseLen
\item param1
\item param2
\item param3
\item param4
\item param5
\item param6
\item param7
\item param8
\item tsSG
\item approved
}

\strong{tagDeps:}
\itemize{
\item tagID
\item deployID
\item status
\item tsStart
\item tsEnd
\item deferSec
\item speciesID
\item markerNumber
\item markerType
\item latitude
\item longitude
\item elevation
\item comments      a JSON-formatted character string of additional properties
\item fullID  tag formatted as
}

\strong{events:}
\itemize{
\item ts    timestamp for event
\item tagID motus tag ID for event
\item event integer: 1 is activation; 0 is deactivation
}


\strong{species:}
\itemize{
\item id  integer species ID
\item english english species name
\item french french species name
\item scientific species name
\item group informal taxonomic group
\item sort taxonomic sorting key
}



\strong{recvDeps:}
\itemize{
\item id   receiver ID ?? how does this differ from motus device ID
\item serno  receiver serial number
\item receiverType
\item deviceID motus device ID
\item macAddress
\item status
\item deployID
\item name
\item fixtureType
\item latitude
\item longitude
\item isMobile
\item tsStart
\item tsEnd
}

\strong{antDeps:}
\itemize{
\item deployID receiver deployment ID
\item port  which port (USB for SGs; BNC for Lotek) the antenna was plugged into
\item antennaType
\item bearing magnetic compass bearing of antenna main axis
\item heightMeters height of antenna above ground
}

\strong{projs:}
\itemize{
\item id motus project ID
\item name motus project name
\item label short project label for graphs
\item tagsPermissions
\item sensorsPermissions
}
}
\description{
Get the motus database of metadata for tags, receivers, projects,
and species.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

