% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamOverrides.R
\name{getParamOverrides}
\alias{getParamOverrides}
\title{Get the parameter overrides for a receiver and time or boot session.}
\usage{
getParamOverrides(serno, monoBN = NULL, tsStart = NA,
  progName = "find_tags_motus", recvDepTol = 3 * 24 * 3600)
}
\arguments{
\item{serno}{character scalar; serial number of receiver, e.g. "Lotek-123" or
"SG-5133BBBK2972"}

\item{monoBN}{integer scalar; boot session for which we want the overrides;
(ignored if \code{serno} is a Lotek receiver}

\item{tsStart}{real scalar; starting timestamp for which we want the overrides;
(ignored if \code{serno} is a sensorgnome)}

\item{progName}{program name for which the parameters are sought;
Default: "find_tags_motus"}

\item{recvDepTol}{numeric scalar; how much slop is allowed when looking
up the project ID for a receiver?  Sometimes, people have recorded incorrect
start of deployment dates, and/or the estimate of boot session time is
inaccurate.  Default: 3*24*3600 which is 3 days.}
}
\value{
a character scalar of parameters, ready for the command line; if there
are no applicable overrides, returns ""
}
\description{
Get the parameter overrides for a receiver and time or boot session.
}
\note{
Overrides come from the paramOverrides table of the motus meta database.
As of 2016 Dec 14, this database is just copied from /sgm/paramOverrides.sqlite, but
eventually motus will provide user-editable fields for these.

An override can be specified for either a whole project, or for a receiver.
In both cases, the override period can be specified as well.  A receiver-specific
override will override a project-wide override, if both are applicable.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

