% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPGID.R
\name{getPGID}
\alias{getPGID}
\title{return the process group ID for the R session}
\usage{
getPGID()
}
\value{
integer scalar giving the process group ID.
}
\description{
This process group ID (PGID) is used as the owner token for
locks acquired on symbols, e.g.  It is known by the shell
scripts which launch the email, status, and process servers,
and is used by those scripts to remove locks if those
servers die.
}
\seealso{
\code{\link{lockSymbol}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
