% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecvModel.R
\name{getRecvModel}
\alias{getRecvModel}
\title{determine the model from a receiver serial number}
\usage{
getRecvModel(serno)
}
\arguments{
\item{serno}{character scalar; receiver serial number, e.g. "Lotek-123"}
}
\value{
a character scalar from this list:

\itemize{
  \item SRX800D
  \item SRX400A
  \item SRX-DL
  \item SRX600
  \item SRX800M/MD
  \item SRX800
  \item BBW
  \item BBBK
  \item RPI2
  \item RPI3
  \item RPIZ
  \item UNKNOWN
}
}
\description{
The serial number determines the receiver model in most cases, but
for a few Lotek receivers, additional fields from a .DTA file are needed.
In those situations, the receiver DB would be examined for appropriate
fields in the meta table.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
