% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecvSrc.R
\name{getRecvSrc}
\alias{getRecvSrc}
\title{Get the src_sqlite for a receiver database given its serial number.}
\usage{
getRecvSrc(serno, dbdir = MOTUS_PATH$RECV, create = TRUE)
}
\arguments{
\item{serno}{receiver serial number}

\item{dbdir}{path to folder with existing receiver databases
Default: \code{MOTUS_PATH$RECV}}

\item{create}{logical scalar; if TRUE, create the database if it doesn't
already exist.}
}
\value{
a src_sqlite for the receiver; if the receiver is new, and \code{create==TRUE} this database
will be empty, but have the correct schema.  If the receiver is new and \code{create==FALSE},
return \code{NULL}.  If \code{serno} does not have the form of a valid serial number,
raise an error.
}
\description{
receiver database files are stored in a single directory, and
have names like "SG-XXXXBBBKYYYY.motus" or "Lotek-NNNNN.motus"
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
