% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSGModel.R
\name{getSGModel}
\alias{getSGModel}
\title{determine the model from a sensorgnome serial number}
\usage{
getSGModel(serno)
}
\arguments{
\item{serno}{character scalar; receiver serial number, e.g. "SG-1234BBBKABCD"}
}
\value{
a character scalar with the receiver model, which so far is one of:
\itemize{
\item BBBK - beaglebone black
\item BBW -  beaglebone white
\item RPI2 - raspberry Pi model 2B
\item RPI3 - raspberry Pi model 3B
\item RPIZ - raspberry Pi model Z wireless
}
}
\description{
The serial number determines the receiver model.  We use the form
SG-XXXXMMMMYYYY where MMMM is the model, and XXXXYYYY are the true
serial number.
}
\details{
the model is stored more or less as-is in the serial number string,
but BBW might have a digit that's not always zero immediately following
the 'BB'.  This function is meant to allow flexibility for future
serial numbers.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
