% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTagProjSrc.R
\name{getTagProjSrc}
\alias{getTagProjSrc}
\title{Get the src_sqlite for a tag project database given its project ID}
\usage{
getTagProjSrc(projectID, dbdir = MOTUS_PATH$TAG_PROJ)
}
\arguments{
\item{projectID}{motus project ID}

\item{dbdir}{path to folder with existing receiver databases
Default: \code{MOTUS_PATH$TAG_PROJ}}
}
\value{
a src_sqlite for the receiver; if the receiver is new, this database
will be empty, but have the correct schema.
}
\description{
tag project database files are stored in a single directory, and
have names like "proj-123.motus"
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
