% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUploadToken.R
\name{getUploadToken}
\alias{getUploadToken}
\title{Get a random token for inclusion in data transfer emails}
\usage{
getUploadToken(user, email, lifeSpan = 14, numBits = 144)
}
\arguments{
\item{user}{name of user on sensorgnome.org}

\item{email}{email address of user on sensorgnome.org}

\item{lifeSpan}{token lifespan, in days.  Default: 14.}

\item{numBites}{number of random bits in token.  Default: 144
The true number of random bits will be somewhat smaller, due
to removal of non-alphnumeric characters '/' and '+'.}
}
\value{
a list with these items:
\itemize{
\item token character scalar giving the token, including its preamble.
\item expiry expiry date, as a timestamp
}
}
\description{
To verify that incoming data are from a legitimate user, and to determine that
user, we ask users to fetch a token from
}
\details{
https://sensorgnome.org/Sending_Data_for_Automatic_Processing

to include in the subject or body of their transfer email.  Tokens
have a fixed lifespan, defaulting to 2 weeks, and up to two
unexpired tokens per user can exist, so that at least one of them
is guaranteed to be good for at least half of the usual lifespan.

Tokens include an 8-character preamble so that they can easily be found
in an email message using a regular expression.

Tokens displayed in a web browser should be selectable by double-clicking
on them, since they use only alphanumeric characters from the ASCII set.

As a side-effect, tokens are stored in the table 'upload_tokens' in
the motus transfer database.  Expired tokens are deleted when new
tokens are generated.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

