% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{gps_for_receiver_project}
\alias{gps_for_receiver_project}
\title{get all GPS fixes from a batch}
\usage{
gps_for_receiver_project(env)
}
\arguments{
\item{projectID}{integer project ID}

\item{batchID}{integer batchID}

\item{ts}{numeric timestamp of latest fix already obtained}

\item{countOnly}{logical return only the count of records}
}
\value{
a data frame with the same schema as the gps table, but JSON-encoded as a list of columns
}
\description{
get all GPS fixes from a batch
}

