% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z%or%.R
\name{\%or\%}
\alias{\%or\%}
\title{perl/javascript style "or"}
\usage{
x \%or\% y
}
\arguments{
\item{x}{first argument, of any class}

\item{y}{second argument, of any class}
}
\value{
\code{x}, unless x is either the logical scalar FALSE, or has
length 0, in which case return \code{y}.  Note that length(NULL) is 0.
}
\description{
returns the first argument, if FALSE or of length 0; otherwise
the second.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
