% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessTagModel.R
\name{guessTagModel}
\alias{guessTagModel}
\title{guess the model of a tag used by the speciesID}
\usage{
guessTagModel(sp)
}
\arguments{
\item{sp}{integer vector of motus speciesID of the tags}
}
\value{
character vector giving guessed tag model e.g. "NTQB-2"
}
\description{
People usually use the largest tag they can reasonably
put on an organism, to maximize the length of time over
which they can get data.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
