% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleArchive.R
\name{handleArchive}
\alias{handleArchive}
\title{Try handle a compressed archive.}
\usage{
handleArchive(path, isdir)
}
\arguments{
\item{path}{the full path to the file with the download link}

\item{isdir}{boolean; TRUE iff the path is a directory}
}
\value{
TRUE iff the \code{path} was an archive and succesfully extracted.
}
\description{
If \code{path} is a compressed archive of known type, it is
extracted into a temporary folder which is then enqueued.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

