% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleAttachments.R
\name{handleAttachments}
\alias{handleAttachments}
\title{Handle attachments to an incoming data email.}
\usage{
handleAttachments(dir)
}
\arguments{
\item{dir}{directory in which the email has been unpacked into parts.}
}
\value{
an integer vector with two elements: the number of
    attachments handled and the total number of message parts.
}
\description{
The email must already have been unpacked into the specified
directory.
}
\details{
Attachments can be of type:
\itemize{
\item .DTA files from a lotek receiver
\item .txt.gz  compressed files from an SG
\item .txt  uncompressed files from an SG
\item .zip compressed archive holding any of the above types
\item .7z ...
\item .rar ...
}

Any attachment whose filename suffix matches one of the above will
be enqueued.  Note:  in the odd case where a user attaches individual
SG .txt and .txt.gz files to a message, these will be processed in
alphabetical order, which should assure the corresponding site
is processed incrementally for each file.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

