% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleDownload.R
\name{handleDownload}
\alias{handleDownload}
\title{Grab content from a downloadable link.}
\usage{
handleDownload(j)
}
\arguments{
\item{j}{the download job; it has these fields:
\itemize{
\item url: download URL
\item method: download method; e.g. googleDrive, wetransferDirect, ...
}

If \code{type="xxx"}, there must exist a handler called \code{downloadXxx};
i.e. the first letter in the type gets capitalized.}
}
\value{
TRUE if the download was successfully handled
}
\description{
The link is contained a file whose name looks like TIMESTAMP_url
The file contains a single line, with this form:
}
\details{
\code{TYPE URL}

where TYPE is the name of a handler, e.g. dropbox for which a function
\code{dowload.TYPE} is defined, and \code{URL} is the location
of the item.
}
\note{
if successful, subjobs to run sanity checks and unpack archives
are enqueued.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{emailServer}}
}

