% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleDownloadableLink.R
\name{handleDownloadableLink}
\alias{handleDownloadableLink}
\title{Grab content from a downloadable link.}
\usage{
handleDownloadableLink(path, isdir, params)
}
\arguments{
\item{path}{the full path to the file with the download link}

\item{isdir}{boolean; TRUE iff the path is a directory}

\item{params}{character vector; not used.}
}
\value{
TRUE if the link was successfully downloaded.
}
\description{
The link is contained a file whose name looks like TIMESTAMP_url
The file contains a single line, with this form:
}
\details{
\code{TYPE URL}

where TYPE is the name of a handler, e.g. dropbox for which a function
\code{dowload.TYPE} is defined, and \code{URL} is the location
of the item.
}
\note{
If the download was successful, the file/folder is enqueued.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

