% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleDownloadableLinks.R
\name{handleDownloadableLinks}
\alias{handleDownloadableLinks}
\title{Parse an email message for links to downloadable content, and handle
these.}
\usage{
handleDownloadableLinks(msg)
}
\arguments{
\item{msg}{character scalar; the message text.}
}
\value{
no return value.
}
\description{
The links must be of the forms specified in \link{\code{dataTransferRegex}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

