% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleDTA.R
\name{handleDTA}
\alias{handleDTA}
\title{handle a folder of .DTA files}
\usage{
handleDTA(path, isdir, params)
}
\arguments{
\item{path}{the full path to the file or directory.  It is only
treated as a file of DTA files if it is a directory whose name
begins with "dta_"}

\item{isdir}{boolean; TRUE iff the path is a directory}

\item{params}{not used}
}
\value{
TRUE iff the .DTA files were successfully handled.
}
\description{
Called by \code{\link{server}} for a file or folder added
to the queue.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

