% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleDTAold.R
\name{handleDTAold}
\alias{handleDTAold}
\title{handle a set of .DTA folders belonging to the same 'old-style' site}
\usage{
handleDTAold(path, isdir, params)
}
\arguments{
\item{path}{the full path to the file or directory.  It is only
treated as a file of DTA files if it is a directory whose name
begins with "dtaold_"}

\item{isdir}{boolean; TRUE iff the path is a directory}

\item{params}{character vector; first item is path to site folder in
old-style hierarchy, with percent-signs representing forward slashes}
}
\value{
TRUE iff the .DTA files for the site were successfully handled.
}
\description{
Called by \code{\link{server}} for a file or folder added
to the queue.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

