% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleEmail.R
\name{handleEmail}
\alias{handleEmail}
\title{handler for incoming email}
\usage{
handleEmail(msg)
}
\arguments{
\item{msg}{the full path to the new message}
}
\value{
TRUE if the file is a valid email; FALSE otherwise
}
\description{
called by \code{\link{emailServer}} for emails.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{emailServer}}
}

