% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleExportData.R
\name{handleExportData}
\alias{handleExportData}
\title{push newly generated data to motus transfer tables}
\usage{
handleExportData(j)
}
\arguments{
\item{j}{the job, with these fields:
\itemize{
\item serno - the receiver serial number
}}
}
\value{
TRUE
}
\description{
Any new batches for this receiver, and their associated runs, hits,
and GPS fixes are added to motus transfer tables.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
