% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleExportPulses.R
\name{handleExportPulses}
\alias{handleExportPulses}
\title{create an interim .sqlite file with tables of pulses and antenna parameters}
\usage{
handleExportPulses(j)
}
\arguments{
\item{j}{the job, with these fields:
\itemize{
\item serno - the receiver serial number
\item batchID - the batchID for which to export pulses
}}
}
\value{
TRUE
}
\description{
For now, beeper tags are supported by generating an .sqlite file with
pulses and parameter settings because the tag finder doesn't yet
support assembling runs of pulses from a beeper tag.
}
\details{
The data are written to tables 'pulses' and 'params' in an sqlite
file.  The sqlite file is stored in the download folder for the project
owning the batch.  This is given by \code{file.path(MOTUS_PATH$WWW,motusProjectID}
The file has a name like `SG-1234BBBK5678_beeper.sqlite`.

Schemas for the tables are copied from the receiver database tables
`params` and `pulses`.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
