% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleFilesWrangled.R
\name{handleFilesWrangled}
\alias{handleFilesWrangled}
\title{handler for completion of email processing}
\usage{
handleFilesWrangled(j)
}
\arguments{
\item{j}{the job}
}
\value{
TRUE if the job was successfully moved to the processing queue
}
\description{
Sends reply to sender giving status of file wrangling, and moves
the job to the top-level motus processing queue, from which one of
the processServer() processes will claim it and process the files
transferred by the email.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{emailServer}}, \code{\link{processServer}}
}

