% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleFullRecvRerun.R
\name{handleFullRecvRerun}
\alias{handleFullRecvRerun}
\title{fully re-run all raw data files from a receiver}
\usage{
handleFullRecvRerun(j)
}
\arguments{
\item{j}{the job with these items:

\itemize{

\item serno character scalar; the receiver serial number

\item [optional] paramOverrides; if present, must be a string scalar
specifying parameters to the tag finder, which will be appended to
any default and registered override values; thus, parameter values specified
here will override those from other sources.  This item is handled
in \code{\link{handleSGfindtags}} or \code{\link{handleLtFindtags}}
}}
}
\value{
TRUE
}
\description{
Called by \code{\link{processServer}}
}
\details{
re-creates the tables in the receiver DB from scratch,
using the files stored in \code{MOTUS_PATH$FILE_REPO/serno},
and then runs the tag finder on all data.
}
\seealso{
\link{\code{processServer}}, \link{\code{handleRerunReceiver}} which reruns
the tag finder on some or all data, but without rebuilding file contents tables
from the file repository.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
