% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleLogs.R
\name{handleLogs}
\alias{handleLogs}
\title{Save a folder of linux system log files from a sensorgnome.}
\usage{
handleLogs(path, isdir)
}
\arguments{
\item{path}{the full path to the file or directory.  It is only
treated as a log file folder if it is a directory whose name
begins with "log_"}

\item{isdir}{boolean; TRUE iff the path is a directory}
}
\value{
TRUE iff the folder could be archived; i.e. iff a valid
    sensorgnome serial number was found in at least one of the log
    files
}
\description{
Called by \code{\link{server}} for a file or folder added
to the queue.
}
\note{

}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

