% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleLtFiles.R
\name{handleLtFiles}
\alias{handleLtFiles}
\title{handle a folder of Lotek .DTA files}
\usage{
handleLtFiles(j)
}
\arguments{
\item{j}{the job with these item(s):
\itemize{
   \item filePath; path to files to be merged; if NULL, defaults to \code{jobPath(j)}
}}
}
\value{
TRUE after queueing jobs
}
\description{
Called by \code{\link{processServer}} for a file or folder added
to the queue.  Merges files into receiver DBs, then queues
a job to run the tag finder on each of these.
}
\note{
if \code{topJob(j)$mergeOnly)} is TRUE, then only merge files
into receiver databases; don't run the tag finder.
}
\seealso{
\link{\code{processServer}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
