% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleNewFiles.R
\name{handleNewFiles}
\alias{handleNewFiles}
\title{handle a new batch of files}
\usage{
handleNewFiles(j)
}
\arguments{
\item{j}{the job}
}
\value{
TRUE; As a side effect, subjobs for handling
various types of known files are queued, like so:

\itemize{

   \item any .DTA (Lotek) files are moved into a new temporary
directory and queued as a subjob of type "DTA"; name collisions
are avoided

   \item any folder containing a file named "syslog(.[0-9](.gz)?)?"
is enqueued as a new subjob of type "logs"

   \item any folder containing a file named "tagreg.txt" is enqueued
as a new subjob of type "registerTags"

   \item any files that don't look like sensorgnome data files,
i.e. that don't have names ending in ".gz" or ".txt.gz", and which
aren't shortened names with a tilde ("~") character are moved into
a new directory and enqueued as a subjob of type "unknownFiles".

   \item any remaining files are assumed to be SG data files, and
a subjob of type "SGfiles" is enqueued to process them.

}
}
\description{
Called by \code{\link{processServer}}.  Any files in subfolders
of the specified job's folder are processed.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{processServer}}
}

