% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlePath.R
\name{handlePath}
\alias{handlePath}
\title{handle a single file or folder}
\usage{
handlePath(path, isdir)
}
\arguments{
\item{path}{the full path to the new file or directory. A single
file is treatedas if it were a directory holding only that file}

\item{isdir}{boolean; TRUE iff the path is a directory}
}
\value{
TRUE iff all files were handled.

The algorithm is this:
\itemize{

   \item any .DTA (Lotek) files are moved into a new temporary
directory whose name ends with _dta, which is then enqueued.

   \item any folder with a file named "syslog" is given a new name
that ends with "_log", and is enqueued

   \item any files with names ending in ".gz" or ".txt.gz", or
which have an 8.3 character filename that includes a tilde ("~")
character are moved into a new temporary directory whose name
ends with _sg, which is then enqueued. See the references
regarding 8.3 filenames

   \item any remaining files of recognized type are enqueued
individually

}
}
\description{
Called by \code{\link{server}} for a file or folder added
to the queue.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\references{
https://en.wikipedia.org/wiki/8.3_filename
}
\seealso{
\code{\link{server}}
}

