% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleQueueArchives.R
\name{handleQueueArchives}
\alias{handleQueueArchives}
\title{Queue a new sub-job for each archive in a folder.}
\usage{
handleQueueArchives(j)
}
\arguments{
\item{j}{the job with item:
\itemize{
\item dir directory in which to search (recursively) for archives
}}
}
\value{
TRUE
}
\description{
For each .7z, .zip, and .rar file, a subjob is queued which
will unpack the archive.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

