% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleRegisterTags.R
\name{handleRegisterTags}
\alias{handleRegisterTags}
\title{handle a batch of tag registrations}
\usage{
handleRegisterTags(j)
}
\arguments{
\item{j}{the job}
}
\value{
TRUE;
}
\description{
Tag registrations are recordings of tag output accompanied by a
text file of metadata.  This function processes the recordings and
pushes the valid ones to motus.org
}
\details{
The job folder must contain a file called tagreg.txt with these lines:

\itemize{
\item  motusProjID:  XXX (numeric project ID)
\item  tagModel: NTQB-??? (as given on the Lotek invoice)
\item  nomFreq: 166.38 (nominal tag frequency, in MHz)
\item  species: XXXXX (optional 4-letter code or motus numeric species ID)
\item  deployDate: YYYY-MM-DD (earliest likely deployment date for any tags)
\item  codeSet: X (optional codeset ; default: 4 for "Lotek4"; can also be 3 for "Lotek3")
}

as well as one or more recording files with names like \code{tagXXX.wav}
where \code{XXX} is the manufacturer's tag ID, typically 1 to 3 digits.
When there are recordings of tags with the same ID but different burst intervals,
the 2nd, 3rd, and so on such tags are given names like \code{tagXXX.1.wav, tagXXX.2.wav, ...}

Called by \code{\link{processServer}}.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{processServer}}, which calls this function.
}

