% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleRerunReceiver.R
\name{handleRerunReceiver}
\alias{handleRerunReceiver}
\title{re-run data from a receiver}
\usage{
handleRerunReceiver(j)
}
\arguments{
\item{j}{the job with these items:

\itemize{

\item serno character scalar; the receiver serial number

\item monoBN integer vector of length 2; the range of receiver boot
sessions to run; If not specified, then for an SG receiver, all
boot sessions are rerun.  Ignored for Lotek receivers.

\item exportOnly logical scalar; if TRUE, skip running the tagfinder

\item cleanup logical scalar; if TRUE, delete all hits, runs,
batches before re-running.  Ignored (and treated as FALSE) if
monoBN is specified.  i.e. cleanup can only be specified for a
full re-run.

}}
}
\value{
TRUE
}
\description{
Called by \code{\link{processServer}}
}
\details{
re-runs the tagfinder for a receiver, possibly limiting the
re-run to one or more boot sessions (for SGs).
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{processServer}}
}

