% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleSG.R
\name{handleSG}
\alias{handleSG}
\title{handle a folder of files from one or more sensorgnomes}
\usage{
handleSG(path, isdir, params)
}
\arguments{
\item{path}{the full path to the file or directory.  It is only
treated as a file of sensorgnome files if it is a directory
whose name ends with "_sg_PATH".}

\item{isdir}{boolean; TRUE iff the path is a directory}

\item{params}{not used}
}
\value{
TRUE iff the sensornome files were successfully handled.
}
\description{
Called by \code{\link{server}} for a file or folder added
to the queue.
}
\details{
For each unique SG serial number found among the names of
    the incoming files, a temporary folder is created, and
    corresponding files are moved there.  That folder is then enqueued
    with a name beggining with "sgsingle_" (see \link{\code{server}})
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

