% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleSGfiles.R
\name{handleSGfiles}
\alias{handleSGfiles}
\title{handle a folder of files from one or more sensorgnomes}
\usage{
handleSGfiles(j)
}
\arguments{
\item{j}{the job}
}
\value{
TRUE
}
\description{
Called by \code{\link{processServer}} for files from a sensorgnome
}
\details{
For each unique SG serial number found among the names of
    the incoming files, queue a subjob for each boot session having
    data files.
}
\note{
if \code{topJob(j)$mergeOnly)} is TRUE, then only merge files
into receiver databases; don't run the tag finder.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{processServer}}
}

