% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleSGnew.R
\name{handleSGnew}
\alias{handleSGnew}
\title{run the tag finder on one sensorgnome boot session, in the new style.}
\usage{
handleSGnew(path, isdir, params)
}
\arguments{
\item{path}{ignored}

\item{isdir}{boolean; TRUE iff the path is a directory; must be FALSE}

\item{params}{character vector of parameters:
\itemize{
\item serno serial number of receiver (including leading 'SG-')
\item monoBN monotonic boot session number
\item boolean TRUE or FALSE; can the tag finder be run with \code{--resume}?
}}
}
\value{
integer vector with these items
\enumerate{
\item batch number
\item number of detections
}
}
\description{
Called by \code{\link{server}}.
}
\details{
Runs the new tag finder for one boot session of a single SG.
The files for that boot session have already been merged into
the DB for that receiver.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{server}}
}

