% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleSGold.R
\name{handleSGold}
\alias{handleSGold}
\title{handle a folder of files from one sensorgnome, in the old style.}
\usage{
handleSGold(path, isdir, params)
}
\arguments{
\item{path}{the full path to the directory of SG files.}

\item{isdir}{boolean; TRUE iff the path is a directory}

\item{params}{character vector; first item is path to site folder in
old-style hierarchy, with percent-signs representing forward slashes}
}
\value{
TRUE iff the sensornome files were successfully handled.
}
\description{
Called by \code{\link{server}} for a file or folder added
to the queue.
}
\details{
Files are moved to the 'incoming' folder of the
appropriate receiver folder of the \code{/SG/} hierarchy.  The
old-style R script \code{update_site.R} is then run for that
site.  This merges files and runs the tag finder on the site's
entire history, against the tag database for that nominal year.
Results are posted to user pages on the sensorgnome.org wiki.
An hour-by-hour tag presence summary and plot are generated,
which compare old and new styles of running the data.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{server}}
}

