% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleUnknownFiles.R
\name{handleUnknownFiles}
\alias{handleUnknownFiles}
\title{handle a set of files of unknown type}
\usage{
handleUnknownFiles(j)
}
\arguments{
\item{j}{the job}
}
\value{
TRUE;
}
\description{
Called by \code{\link{processServer}}.  Files are
retained, and the sender is emailed a list of the files.
}
\seealso{
\code{\link{processServer}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
