% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleUnpackArchive.R
\name{handleUnpackArchive}
\alias{handleUnpackArchive}
\title{Try unpack a compressed archive into its job folder.}
\usage{
handleUnpackArchive(j)
}
\arguments{
\item{j}{the job, which has this parameter:
\itemize{
\item file: the path to the archive file
}}
}
\value{
TRUE on success; FALSE otherwise
}
\description{
Try unpack a compressed archive into its job folder.
}
\note{
if successfull, new subjobs to run sanity checks and unpack nested
archives are queued.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
