% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleUnvalidatedEmail.R
\name{handleUnvalidatedEmail}
\alias{handleUnvalidatedEmail}
\title{handler for email with missing or expired validation token}
\usage{
handleUnvalidatedEmail(j)
}
\arguments{
\item{j}{the job}
}
\value{
TRUE if the unvalidated email message was successfully handled
}
\description{
Sends a copy of the unvalidated email to MOTUS_ADMIN_EMAIL, which
allows that person to paste in a token and resubmit the email if
it is valid.
}
\details{
Called by \code{\link{emailServer}}
}
\seealso{
\code{\link{emailServer}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
