% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleUploadFile.R
\name{handleUploadFile}
\alias{handleUploadFile}
\title{handle an uploaded file}
\usage{
handleUploadFile(j)
}
\arguments{
\item{j}{the job}
}
\value{
TRUE; As a side effect, subjobs for sanity checking files and
unpacking archives are queued.
}
\description{
Called by \code{\link{processServer}}.  The file is sanity checked,
unpacked if necessary, then run.
}
\seealso{
\code{\link{processServer}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
