% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{hits_for_receiver_project}
\alias{hits_for_receiver_project}
\title{get hits by receiver project from a batch (i.e. all hits from the batch)}
\usage{
hits_for_receiver_project(env)
}
\arguments{
\item{projectID}{integer project ID}

\item{batchID}{integer batchID}

\item{hitID}{integer ID of largest hit already obtained}

\item{countOnly}{logical return only the count of records}
}
\value{
a data frame with the same schema as the hits table, but JSON-encoded as a list of columns
}
\description{
get hits by receiver project from a batch (i.e. all hits from the batch)
}

