% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{hits_for_tag_project}
\alias{hits_for_tag_project}
\title{get hits by tag project from a batch}
\usage{
hits_for_tag_project(env)
}
\arguments{
\item{projectID}{integer project ID}

\item{batchID}{integer batchID}

\item{hitID}{integer ID of largest hit already obtained}
}
\value{
a data frame with the same schema as the hits table, but JSON-encoded as a list of columns
}
\description{
get hits by tag project from a batch
}
