% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgSyncDBtoRepo.R
\name{infimum}
\alias{infimum}
\title{get the greatest lower bound of each number in one set from the other}
\usage{
infimum(s1, s2)
}
\arguments{
\item{s1}{set of numbers to find lower bounds for}

\item{s2}{set of numbers among which to find the greatest lower bounds.}
}
\value{
vector rv with same type as s2 and length as s1, with
rv[i] = max(s2[s2 <= s1[i]]), if any(s2 <= s1[i]);
        or NA, if all(s2 > s1[i])
i.e. for each number in s1, its greatest lower bound among the numbers in s2,
or NA if none exists.
}
\description{
get the greatest lower bound of each number in one set from the other
}
\details{
the algorithm run time is O(n log(n)), where n = max(length(s1), length(s2))
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
