% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTopJob.R
\name{isTopJob}
\alias{isTopJob}
\title{is this a top-level job?}
\usage{
isTopJob(j)
}
\arguments{
\item{j}{Twig object representing the job}
}
\value{
TRUE if and only if the job does not have a parent job.
}
\description{
is this a top-level job?
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
