% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTopJobDone.R
\name{isTopJobDone}
\alias{isTopJobDone}
\title{Check whether the top job of which a job is a subjob is done;
i.e. whether all subjobs completed, whether successfully or not.}
\usage{
isTopJobDone(j)
}
\arguments{
\item{j}{the job}
}
\value{
TRUE iff all other jobs having the same stump have \code{done != 0}
}
\description{
Check whether the top job of which a job is a subjob is done;
i.e. whether all subjobs completed, whether successfully or not.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

