% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobAtPath.R
\name{jobAtPath}
\alias{jobAtPath}
\title{Return the \code{motusJob} object corresponding to a path.
An entry in the server database is created.}
\usage{
jobAtPath(path)
}
\arguments{
\item{path}{location of job folder.}
}
\value{
This function returns an object of class "motusJob".  The
    object is a named integer scalar whose value is the job number,
    and whose name is the full path to the new folder.
If \code{path} does not refer to a valid job, then return NULL.
}
\description{
Return the \code{motusJob} object corresponding to a path.
An entry in the server database is created.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

