% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobFail.R
\name{jobFail}
\alias{jobFail}
\title{mark a job as failed}
\usage{
jobFail(j, code = -1, ...)
}
\arguments{
\item{j}{the job}

\item{...}{character vector of messages to add to job's log}

\item{code;}{the value to store in "done"; should be non-zero
and defaults to -1.}
}
\value{
no return value
}
\description{
record an error message for the job, mark it with done=-1,
and move its folder to MOTUS_PATH$ERRORS
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

