% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobHasFolder.R
\name{jobHasFolder}
\alias{jobHasFolder}
\title{does a job have a folder in the filesystem?}
\usage{
jobHasFolder(j)
}
\arguments{
\item{j}{Twig object representing the job}
}
\value{
TRUE if and only if the job 'should' have a folder in the
    filesystem.  This is indicated by a non-null value for the
    job's \code{path} column in the Copse database.
}
\description{
does a job have a folder in the filesystem?
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

