% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobLog.R
\name{jobLog}
\alias{jobLog}
\title{Add a message to the log for a job.  Log messages are saved in a
JSON field called "log" in the top-level ancestor of the job.}
\usage{
jobLog(j, msg)
}
\arguments{
\item{j}{the job}

\item{msg}{character vector of messages

The messages in \code{msg} are joined with "\\n" and appended to any already existing for this job.}
}
\description{
Add a message to the log for a job.  Log messages are saved in a
JSON field called "log" in the top-level ancestor of the job.
}

