% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobPath.R
\name{jobPath}
\alias{jobPath}
\title{get the filesystem path for a job}
\usage{
jobPath(j)
}
\arguments{
\item{j}{the job}
}
\value{
A character scalar giving the full path to the job's folder, or
to the most recent ancestor's folder if this job doesn't have one.
To test whether a job has a folder, use \link{\code{jobHasFolder()}}
}
\description{
get the filesystem path for a job
}
\details{
If a job has a filesystem folder, its path is a non-null character string
giving the path to that folder relative to the path to its parent job.

If the job does not have a filesystem folder, this function returns the
path of the most recent ancestor which does have a filesystem folder.

This function uses an SQLite recursive common table entity query to
generate the path on the fly from the \code{$path} components of
this jobs and its ancestors.  This traces the job up to its top job,
pre-pending non-null path components as it goes.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{Copse}}
}

