% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightWeightCopy.R
\name{lightWeightCopy}
\alias{lightWeightCopy}
\title{Make a light-weight recursive copy of a folder.}
\usage{
lightWeightCopy(src, dst)
}
\arguments{
\item{src}{character vector; paths to source files / folders}

\item{dst}{character scalar; path to target folder}
}
\value{
TRUE on success, FALSE otherwise.
}
\description{
This uses hardlinks wherever possible, so that file
data are not copied, but falls back to normal copy where
necessary.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

