% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadJobs.R
\name{loadJobs}
\alias{loadJobs}
\title{load existing jobs from server database}
\usage{
loadJobs()
}
\value{
no return value.
}
\description{
A global JOBS object is created, which manages jobs.
It is populated from the "jobs" table in the server database.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

