% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockReceiver.R
\name{lockReceiver}
\alias{lockReceiver}
\title{try to lock or unlock a receiver, preventing any other processServer() processes from
using it.}
\usage{
lockReceiver(serno, lock = TRUE)
}
\arguments{
\item{serno}{receiver serial number}

\item{lock}{logical scalar; if TRUE, try unlock the receiver; otherwise, release
any lock.}
}
\value{
if \code{lock} is \code{TRUE}, then return TRUE if this process now has an exclusive lock on the receiver, otherwise FALSE.
If \code{lock} is \code{FALSE}, then always return TRUE.
}
\description{
This prevents interleaved access while running the tag finder for a receiver.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

