% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockSymbol.R
\name{lockSymbol}
\alias{lockSymbol}
\title{Try to lock or unlock a symbol, preventing any other cooperating motusServer process
from using it.}
\usage{
lockSymbol(symbol, owner = getPGID(), lock = TRUE)
}
\arguments{
\item{symbol}{character scalar; this can be a receiver serial
number, database filename, or other arbitrary unique symbol.}

\item{owner}{integer; defaults to process group id of the R
session, but any integer uniquely associated with the process
can be used. The owner is recorded in the symLocks table of the
motus server database when locking is successful (indeed,
'successful' means that after attempting to lock the symbol,
the owner associated with it is this parameter).}

\item{lock}{logical scalar; if TRUE, try unlock the receiver;
otherwise, release any lock.}
}
\value{
if \code{lock} is \code{TRUE}, then return TRUE if this
    process now has an exclusive lock on the symbol, otherwise
    FALSE.  If \code{lock} is \code{FALSE}, then always return
    TRUE.
}
\description{
This is mainly to prevent interleaved access to receiver and other databases.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

