% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltDefaultFindTagsParams.R
\docType{data}
\name{ltDefaultFindTagsParams}
\alias{ltDefaultFindTagsParams}
\title{default list of parameters for the tag finder for Lotek data}
\format{An object of class \code{character} of length 1.}
\usage{
ltDefaultFindTagsParams
}
\description{
These are:
\itemize{
\item   pulses_to_confirm      number of pulses required from a tag before its identity is treated as confirmed
\item   pulse_slop             ms; maximum allowed slop in pulse timing within a burst
\item   burst_slop             ms; maximum allowed slop for a burst interval
\item   burst_slop_expansion   ms; amount by which burst slop is allowed to grow per missed burst
\item   use_events             use the events table in the tag database, so that tags are only sought while alive
\item   max_skipped_bursts     seconds; maximum amount of time between detections in a tag run
\item   default_freq           MHz; default antenna frequency, when not provided in input
\item   sig_slop               signal slop; not supposed to affect data from Lotek receivers, but set to negative
                               as a workaround for https://github.com/jbrzusto/find_tags/issues/74
}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\keyword{datasets}
