% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltDTAregex.R
\docType{data}
\name{ltDTAregex}
\alias{ltDTAregex}
\title{A PCRE-style regular expression for matching interesting parts of
Lotek .DTA files.  Note that we use the 'possessive' version of
non-zero repeats  i.e. '++' instead of '+', as this prevents huge
consumption of stack space for possible backtracking, which we never
need here.}
\format{An object of class \code{character} of length 1.}
\usage{
ltDTAregex
}
\description{
A PCRE-style regular expression for matching interesting parts of
Lotek .DTA files.  Note that we use the 'possessive' version of
non-zero repeats  i.e. '++' instead of '+', as this prevents huge
consumption of stack space for possible backtracking, which we never
need here.
}
\keyword{datasets}

