% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltFindTags.R
\name{ltFindTags}
\alias{ltFindTags}
\title{find tags in a Lotek receiver dataset}
\usage{
ltFindTags(src, tagDB, par = ltDefaultFindTagsParams, toFile = NULL,
  keepOld = 0)
}
\arguments{
\item{src}{dplyr src_sqlite to (lotek) receiver database}

\item{tagDB}{path to sqlite tag registration database}

\item{par}{list of parameters to the filtertags code.  Defaults
to \link{\code{ltDefaultFindTagsParams}}}

\item{toFile}{if not NULL (the default), write lotek output lines
to the this file, rather than passing them to the tag finder.
No tag finding is performed in this case.}

\item{keepOld}{the maximum number of batches of output from previous
runs to keep in the database.  Oldest are deleted first.
If < 0, do note delete any previous runs.}
}
\value{
the number of tag detections in the stream.
}
\description{
This function searches for sequences of tag detections
corresponding to registered (ID, burst interval) pairs and adds
them to the hits, runs, batches etc. tables in the receiver
database.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

