% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltGetCodeset.R
\name{ltGetCodeset}
\alias{ltGetCodeset}
\title{Return the database of canonical parameters for a Lotek codeset.}
\usage{
ltGetCodeset(codeSet = c("Lotek4", "Lotek3"), pathOnly = FALSE)
}
\arguments{
\item{codeSet:}{character scalar; only "Lotek3" or "Lotek4"
are permitted so far.}

\item{pathOnly:}{logical scalar; if TRUE, return only the path
to the .sqlite database containing the codeset.  Default: FALSE}
}
\value{
a dplyr data_frame with these columns:

\itemize{
\item id tag ID
\item g1 gap between 1st, 2nd pulses, in ms
\item g2 gap between 2nd, 3rd pulses, in ms
\item g3 gap between 3rd, 4th pulses, in ms
}
}
\description{
This function returns a dplyr data_frame with the nominal gap
values for all Lotek tag IDs in the specified codeset.
}
\note{
This function will only work on the sensorgnome.org server,
    where we have Lotek's permission to host an encrypted copy of
    their ID code database.  At server boot time, a user with sudo
    privileges must run a script to decrypt the database into
    locked system memory, from where it can only be accessed by
    user "sg" (or, of course, those with sudo privileges).  The
    decrypted versions reside in

    /home/sg/ramfs/Lotek4.sqlite and
    /home/sg/ramfs/Lotek3.sqlite

    while the encrypted versions are in /home/sg/lotekdb

    The decryption script is in /home/sg/bin/decryptLotekDB.R
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
