% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltMergeFiles.R
\name{ltMergeFiles}
\alias{ltMergeFiles}
\title{Merge a batch of Lotek .DTA files with an existing database.}
\usage{
ltMergeFiles(files, j, dbdir = MOTUS_PATH$RECV)
}
\arguments{
\item{files}{either a character vector of full paths to files, or the full
path to a directory, which will be searched recursively for .DTA files.}

\item{j}{job, whose ID will be recorded for any new / changed DTAfile record}

\item{dbdir}{path to folder with existing receiver databases; defaults
to \code{MOTUS_PATH$RECV}}
}
\value{
a data_frame reporting the fate of each file, with these columns:
\enumerate{
\item fullname - full path to filename
\item nameNew  - TRUE iff file of this name was not yet in database
\item dataNew  - TRUE iff file with these contents was not yet in database
\item use      - TRUE iff file copied to database
\item err      - if not NA, any error message generated by trying to parse the file
\item serno    - serial number (e.g. "Lotek-1234") of receiver this file is from
\item ts       - first timestamp from a detection in this file
\item tsLast   - last timestamp from a detection in this file
\item inRepo   - file came from file_repo (rather than being a new upload; happens when fully reprocessing)
\item fid      - fileID, if file in or added to DB
}
}
\description{
Merge a batch of Lotek .DTA files with an existing database.
}
\details{
Any files which are not identical to an existing file are saved
in the file repository at \code{MOTUS_PATH$FILE_REPO}, in addition to
being stored internally in the receiver database.  The names will be
modified if necessary to avoid collisions with existing files.

Remaining files are deleted.
}
\note{
If err is not NA for a file, then other fields for that file
    might not be set appropriately in the return value.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
