% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltRunNewFiles.R
\name{ltRunNewFiles}
\alias{ltRunNewFiles}
\title{Process a new batch of files from a Lotek receiver.}
\usage{
ltRunNewFiles(files, dbdir = MOTUS_PATH$RECV, ...)
}
\arguments{
\item{files}{either a character vector of full paths to files, or
the full path to a directory, which will be searched
recursively for raw sensorgnome data files.}

\item{dbdir}{path to folder with existing receiver databases
Default: \code{MOTUS_PATH$RECV}}

\item{...}{additional parameters to the tag finder; see \link{\code{ltFindTags}}}
}
\value{
the data.frame returned by \code{ltMergeFiles(files, dbdir)}.
}
\description{
Merge the files into the appropriate receiver databases, and run the tag
finder on each receiver.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{ltMergeFiles} and \code{ltFindTags}, which this function calls.
}

