% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltSyncDBtoRepo.R
\name{ltSyncDBtoRepo}
\alias{ltSyncDBtoRepo}
\title{Make sure all files stored internally in a Lotek receiver DB
are also present in the file_repo.}
\usage{
ltSyncDBtoRepo(serno, dbdir = MOTUS_PATH$RECV, repo = MOTUS_PATH$FILE_REPO,
  bkup = MOTUS_PATH$TRASH)
}
\arguments{
\item{serno}{receiver serial number}

\item{dbdir}{path to folder with existing receiver databases
Default: \code{MOTUS_PATH$RECV}}

\item{repo}{path to folder with existing receiver file repos
Default: \code{MOTUS_PATH$FILE_REPO}}

\item{bkup}{path to folder for storing replaced files as backup.  They will
be stored in a folder whose name is the receiver serial number
Default: \code{MOTUS_PATH$TRASH}}
}
\value{
a data.frame with these columns:
\itemize{
\item name - character; bare filename, without compression extension
\item status - integer; status of DB file; possible values:
\itemize{
\item 0: file already in repo and contents there of same size equal or larger than DB contents, so no action taken
\item 1: file already in repo but contents there were smaller than DB contents, so file replaced
and old copy sent to \code{oldrepo}
\item 2: file not in repo, so added to repo
}
}

Returns NULL if no valid data files were found.
}
\description{
Make sure all files stored internally in a Lotek receiver DB
are also present in the file_repo.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
