% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{makeBody}
\alias{makeBody}
\title{make the body for a reply
converts a list or data.frame to json using toJSON with options:
\itemize{
\item auto_unbox=TRUE
\item dataframe="columns"
}
then compresses the result
with memCompress and method "bzip2"}
\usage{
makeBody(x)
}
\arguments{
\item{x}{list or data.frame to encode and compress}
}
\value{
a raw vector
}
\description{
make the body for a reply
converts a list or data.frame to json using toJSON with options:
\itemize{
\item auto_unbox=TRUE
\item dataframe="columns"
}
then compresses the result
with memCompress and method "bzip2"
}
