% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueuePath.R
\name{makeQueuePath}
\alias{makeQueuePath}
\title{Create a filename or folder for the motus processing queue.}
\usage{
makeQueuePath(..., isdir = TRUE, dir = MOTUS_PATH$TMP, create = TRUE)
}
\arguments{
\item{...}{additional components for the file or folder name}

\item{isdir}{boolean scalar: if TRUE, the default, a temporary
folder is created; otherwise, a file is created}

\item{dir}{path to folder in which to create new folder or file;
default: \code{MOTUS_PATH$TMP}}

\item{create}{if TRUE (the default), create the folder or (empty)
file before returning.}
}
\value{
the path to the new file or folder
}
\description{
Create a filename or folder for the motus processing queue.
}
\details{
The name for the new path is a timestamp with any additional
components specified in \code{...} pasted afterward, separated by MOTUS_QUEUE_SEP
which defaults to ',' (comma); e.g. 2016-09-14T00-43-29.586886,dta
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

