% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReceiverPlot.R
\name{makeReceiverPlot}
\alias{makeReceiverPlot}
\title{make a plot of tag detections and status of antennas}
\usage{
makeReceiverPlot(recv, meta = NULL, title = "", condense = 3600,
  ts = NULL, monoBN = NULL)
}
\arguments{
\item{recv}{path or dplyr::src_sqlite to the receiver database}

\item{meta}{path or dplyr::src_sqlite to the file with metadata; if
NULL (the default), assume metadata tables are in the database
given by \code{recv}}

\item{title}{additional title to identify plot, beyond receiver
serial number}

\item{condense}{double scalar; if not NULL, specifies the
condensation period: only the first detection of a given tag on
a given antenna per condensation period is plotted.  Default:
3600, meaning at most one detection of each tag on each antenna
is shown per hour, and it will be the first for that hour.}

\item{ts}{double vector of length 2; range of timestamps to plot;
Default: NULL, meaning no restriction on timestamps.}

\item{monoBN}{integer vector of length 2; range of boot sessions to
plot; Default: NULL, meaning no restriction on boot sessions.}
}
\value{
a list with these items:
\itemize{

\item width: recommended plot width, in pixels

\item height: recommended plot height, in pixels

\item plot: object of class \code{trellis}; to generate a plot from
this:

\itemize{

\item open a graphics device, using the recommended width and
height

\item print the plot

\item close the device

}

\item data: dataframe of the data plotted, with these columns

\itemize{

\item ant: antenna number

\item fullID full tag ID, or " Antenna N Status"; the latter
'detections' just indicate the antenna was functioning in the hour
centred on the timestamp.  Also, a fullID of
" Reboot Odometer" indicate the approximate timestamps at
which the receiver rebooted, and the \code{ant} field for these
records is the last digit of the boot session count.


\item bin: the bin number for condensation

\item ts: the timestamp at the start of the condensation bin

\item n: the number of detections (will be at least 1)

\item freq: the mean offset frequency of detections

\item sig: the \emph{maximum} signal strength

}

Note: \code{n}, \code{freq}, and \code{sig} are for the given tag and antenna
during the condensation period.

}
}
\description{
The plot is an object of class \code{trellis} which can then
be sent to an open plotting device using \code{print()}.
Note: only detections with valid timestamps, i.e. after
1 Jan 2010, are shown.
}
\details{
If both \code{ts} and \code{monoBN} are NULL, then all
    detections in database \code{recv} are plotted.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

