% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTagProjDB.R
\name{makeTagProjDB}
\alias{makeTagProjDB}
\title{create an sqlite database of all detections of tags from a motus project and
the associated metadata for both tags and receivers}
\usage{
makeTagProjDB(projectID = 83, maxHits = NULL)
}
\arguments{
\item{projectID}{integer scalar motus project ID}

\item{maxHits}{if not NULL, specifies a maximum number of tag hits returned;
only intended for testing}
}
\value{
returns a src_sqlite to the SQLite database
which will be called proj-NNN.motus and be in folder \code{MOTUS_PATH$TAG_PROJ}
}
\description{
create an sqlite database of all detections of tags from a motus project and
the associated metadata for both tags and receivers
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

