% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusServer2.R
\name{makeWhere}
\alias{makeWhere}
\title{add condition(s) to the where clause; strip
any leading "where " from the items in clauses,
then join them together using conj and add a preceding "where "}
\usage{
makeWhere(clauses, conj = "and", prep = "where")
}
\description{
add condition(s) to the where clause; strip
any leading "where " from the items in clauses,
then join them together using conj and add a preceding "where "
}
