% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchTags.R
\name{matchTags}
\alias{matchTags}
\title{find which tags match registered tags}
\usage{
matchTags(u, m, id, bi, yr, bi.digits = 0, id.digits = 0)
}
\arguments{
\item{u:}{user tags; a dplyr::tbl_df object, having at least the
columns representing manufacturer tag ID, burst interval, and
registration year}

\item{m:}{motus tags; a dplyr::tbl_df object with one
motus-registered tag per row.  Typically, this will be the
result of calling. \code{getProjectTags()}}

\item{id:}{name of column in \code{u} representing the tags'
manufacturer ID.}

\item{bi:}{name of column in \code{u} representing burst interval,
in seconds.}

\item{yr:}{name of column in \code{u} representing tag registration
year.}

\item{bi.digits:}{number of digits after the decimal point to use
in matching burst interval.  Burst intervals are rounded to
this number of digits before matching.  Defaults to 0, meaning
integer values are used.}

\item{id.digits:}{number of digits after the decimal point to use
in matching manufacturer tag ID.  Tags might have been
registered with digits after the decimal point to represent
duplicate IDs distinguished by burst interval in the same year
and project.  IDs are rounded to this number of digits before
matching. Defaults to 0, meaning integer values are used.}
}
\value{
a list with four items:
\enumerate{
\item  unique: the join of rows from \code{u} with their unique match in \code{m}, for those
user tags with a unique match in the motus tag list

\item multi: the join of rows from \code{u} with each of their matchs in \code{m}, for those
user tags having multiple matches in the motus tag list

\item collide: the join of rows from \code{u} with each of their matchs in \code{m}, for those
motus tags having multiple matches in the user tag list; these are collisions
in the user tag list.

\item none: the rows from \code{u} having no match in \code{m}

}
}
\description{
A set of user tags is looked up in a set of motus-registered tags.
The user tag set is divided into three subsets: tags with unique, multiple,
or no matches in the motus set, respectively.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
