% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processServer.R
\name{maybeRetireJob}
\alias{maybeRetireJob}
\title{if a job is done and none of its subjobs are left in the queue, then move
it to the 'done' folder}
\usage{
maybeRetireJob(j)
}
\arguments{
\item{j}{a job}
}
\value{
TRUE if the job is being retired; FALSE otherwise
}
\description{
if a job is done and none of its subjobs are left in the queue, then move
it to the 'done' folder
}
