% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{metadata_for_receivers}
\alias{metadata_for_receivers}
\title{get metadata for receivers}
\usage{
metadata_for_receivers(env)
}
\arguments{
\item{deviceIDs;}{integer vector of motus device IDs; receiver
metadata will only be returned for receivers whose project has
indicated their metadata are public, or receivers in one of the
projects the user has permissions to.}
}
\value{
a list with these items:
\itemize{
   \item recvDeps; a list with these vector items:
   \itemize{
      \item deployID; integer deployment ID (internal to motus, but links to antDeps)
      \item projectID; integer ID of project that deployed the receiver
      \item serno; character serial number, e.g. "SG-1214BBBK3999", "Lotek-8681"
      \item receiverType; character "SENSORGNOME" or "LOTEK"
      \item deviceID; integer device ID (internal to motus)
      \item status; character deployment status
      \item name; character; typically a site name
      \item fixtureType; character; what is the receiver mounted on?
      \item latitude; numeric (initial) location, degrees North
      \item longitude; numeric (initial) location, degrees East
      \item elevation; numeric (initial) location, metres ASL
      \item isMobile; integer non-zero means a mobile deployment
      \item tsStart; numeric; timestamp of deployment start
      \item tsEnd; numeric; timestamp of deployment end, or NA if ongoing
   }
   \item antDeps; a list with these vector items:
   \itemize{
      \item deployID; integer, links to deployID in recvDeps table
      \item port; integer, which receiver port (USB for SGs, BNC for Lotek) the antenna is connected to
      \item antennaType; character; e.g. "Yagi-5", "omni"
      \item bearing; numeric compass angle at which antenna is pointing; degrees clockwise from magnetic north
      \item heightMeters; numeric height of main antenna element above ground
      \item cableLengthMeters; numeric length of coaxial cable from antenna to receiver, in metres
      \item cableType: character; type of cable; e.g. "RG-58"
      \item mountDistanceMeters; numeric distance of mounting point from receiver, in metres
      \item mountBearing; numeric compass angle from receiver to antenna mount; degrees clockwise from magnetic north
      \item polarization2; numeric angle giving tilt from "normal" position, in degrees
      \item polarization1; numeric angle giving rotation of antenna about own axis, in degrees.
   }
}
}
\description{
get metadata for receivers
}
\note{
only metadata which are public, or which are from projects
    the user has permission to are returned.
}
