% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{metadata_for_tags}
\alias{metadata_for_tags}
\title{get metadata for tags}
\usage{
metadata_for_tags(env)
}
\arguments{
\item{projectID}{integer project ID}

\item{motusTagIDs}{integer vector of tag IDs for which metadata are sought}
}
\value{
a data frame with the same schema as the batches table, but JSON-encoded as a list of columns
}
\description{
get metadata for tags
}

