% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusConstants.R
\docType{data}
\name{MOTUS_API_ENTRY_POINTS}
\alias{MOTUS_API_ENTRY_POINTS}
\title{Constants for the motus package.}
\format{An object of class \code{character} of length 1.}
\usage{
MOTUS_API_ENTRY_POINTS
}
\description{
API entry points:
}
\keyword{datasets}

