% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusDeployTag.R
\name{motusDeployTag}
\alias{motusDeployTag}
\title{register a tag deployment with motus}
\usage{
motusDeployTag(tagID, projectID, status = c("pending", "deploy", "terminate"),
  tsStart = NULL, tsEnd = NULL, deferTime = 0, speciesID = NULL,
  markerType = NULL, markerNumber = NULL, lat = NULL, lon = NULL,
  elev = NULL, comments = NULL, properties = NULL,
  ts = as.numeric(Sys.time()))
}
\arguments{
\item{tagID}{Unique numeric ID assigned to the tag, provided at the
time of tag registration.}

\item{projectID}{numeric ID of motus project for which this tag is
being deployed.}

\item{status}{Status from one of the possible following values:
pending, deploy, terminate. "pending" indicates that the
deployment record isn’t ready to be finalized for deployment
yet. "deploy" indicates that the deployment can be
activated. "terminate" will attempt to finalize a deployment by
providing an end date. Note that deployments can only be
activated or terminated if all required information has been
provided. Attempts to deploy or terminate an incomplete
deployment will return an error. At the moment, several of the
parameters can no longer be modified once a deployment is
activated, and the status cannot be changed back to an earlier
status (permitted sequence is pending -> deploy -> terminate,
but levels can be omitted).}

\item{tsStart}{[optional] Timestamp for start of deployment. If the tag has a
deferred time lag (deferTime), this is the time at which the
tag is activated, so that the time when the tag is
expected to be active is given by tsStart + deferTime.}

\item{tsEnd}{[optional] Timestamp for end of deployment. Required
for status=terminate.}

\item{deferTime}{[optional] Defer time (in seconds from
tsStart). Some tags are capable of deferred activation - they
don't start transmitting until some (possibly large) number of
seconds after activation.}

\item{speciesID}{[optional] Numeric ID (integer) of the species on
which the tag is being deployed. You can obtain this value by
using \code{motusListSpecies()} to search by name or code.
Alternatively, this can be a character scalar giving the species
4-lettercode; e.g. "SESA"}

\item{markerType}{[optional] Type of marker
(e.g. "metal band”, "color band”)}

\item{markerNumber}{[optional] Marker number or descriptor
(e.g. "1234-56789” or "L:Red/Blue,R:Metal 1234-56789”)}

\item{lat}{[optional] Latitude (decimal degrees) of the deployment
site. E.g. 45.123.}

\item{lon}{[optional] Longitude (decimal degrees) of the deployment
site. E.g. -60.325.}

\item{elev}{[optional] Elevation above sea level (meters) of the
deployment site. E.g. 23.}

\item{comments}{[optional] deployment-related comment}

\item{properties}{[optional] list or vector; additional properties
of the deployment or the organism on which the tag is deployed.
This will be formatted as a JSON string then inserted into the
\code{properties} field of the database.  FIXME: for now,
it goes into the comments field.}

\item{ts}{[optional] Time at which deployment information
was generated.  Defaults to time at which function is called.}
}
\description{
register a tag deployment with motus
}
\note{
All timestamps must be in the form returned by \code{Sys.time()}
and by the \code{ymd(), ymd_hms(),..} functions from the \code{lubridate}
package.  i.e. they are numbers representing seconds elapsed since the start of
1 Jan. 1970, GMT.

For timestamps in the future, this function uses the \code{tsAnticipatedStart} parameter
to the motus /tags/deploy API.  For timestamps in the past, it uses \code{tsStart}.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

