% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusListProjects.R
\name{motusListProjects}
\alias{motusListProjects}
\title{return a list of all motus projects}
\usage{
motusListProjects(type = c("both", "tag", "sensor"), ...)
}
\arguments{
\item{type:}{"tag", "sensor", or "both" (the default) Specifies the
type of project to be listed.}

\item{...:}{additional parameters to motusQuery()}
}
\value{
the list of motus projects, a data.frame with these columns

\itemize{
\item id integer; project ID
\item name character; long name
\item label character; short name for plotting etc.
\item tagsPermissions; integer
\item sensorsPermissions; integer
}
}
\description{
return a list of all motus projects
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
