% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusListSensorDeps.R
\name{motusListSensorDeps}
\alias{motusListSensorDeps}
\title{return a list motus sensor deployments for a project}
\usage{
motusListSensorDeps(projectID = NULL, ...)
}
\arguments{
\item{projectID:}{integer; motus internal project ID}

\item{...:}{additional parameters to motusQuery()}
}
\value{
the list of sensor deployments for this project
}
\description{
return a list motus sensor deployments for a project
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
