% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusListSensors.R
\name{motusListSensors}
\alias{motusListSensors}
\title{return a list of all motus sensors}
\usage{
motusListSensors(projectID = NULL, year = NULL, serialNo = NULL, ...)
}
\arguments{
\item{projectID:}{integer; motus internal project ID}

\item{year:}{integer; year of sensor deployment}

\item{serialNo:}{character; serial "number" for sensor, if a
specific sensor is sought.}

\item{...:}{additional parameters to motusQuery()}
}
\value{
the list of motus sensors.
}
\description{
return a list of all motus sensors
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
