% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusListTags.R
\name{motusListTags}
\alias{motusListTags}
\title{return a list of all motus tags}
\usage{
motusListTags(projectID, year = NULL, mfgID = NULL, ...)
}
\arguments{
\item{projectID:}{integer scalar; motus internal project ID}

\item{year:}{integer scalar; year of tag registration}

\item{mfgID:}{character scalar; typically a small integer.}

\item{...:}{additional parameters to motusQuery()}
}
\value{
the list of motus tags registered in the given year to the
    given project.
}
\description{
return a list of all motus tags
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

