% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusLoadSecrets.R
\name{motusLoadSecrets}
\alias{motusLoadSecrets}
\title{load secrets for motus web API}
\usage{
motusLoadSecrets(f, quiet = FALSE)
}
\arguments{
\item{f}{filename: path to json-formatted file of secrets,
with these items:
\enumerate{
  \item key: API key; a secret
  \item user: username at motus-wts.org
  \item passwd: password at motus-wts.org
  \item serno: serial number of a sensorgnome
}}

\item{quiet}{if TRUE, return silently on failure; otherwise,
report errors.}
}
\value{
TRUE if secrets were successfully loaded.
}
\description{
load secrets for motus web API
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

