% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusQuery.R
\name{motusQuery}
\alias{motusQuery}
\title{Send a query to the motus API}
\usage{
motusQuery(API, params = NULL, requestType = "post", show = FALSE,
  json = FALSE, serno = NULL, masterKey = NULL, ...)
}
\arguments{
\item{API}{one of the MOTUS_API_... constants}

\item{params}{named list of API-specific parameters}

\item{requestType}{"post" or "get"}

\item{show}{if TRUE, print the request to the console before submitting to motus}

\item{json}{if TRUE, return results as JSON-format string; otherwise, as R list}

\item{serno}{serial number of receiver from which request is being
sent; if NULL, the default, uses MOTUS_SECRETS$serno.}

\item{masterKey}{if NULL (the default), use key from the
MOTUS_SECRETS object.  Otherwise, \code{masterKey} is the name
of a file to read the secret key from.}

\item{...}{if present, additional CURL options; these are guaranteed to
override any values set by this function's code.}
}
\value{
the result of sending the request to the motus API.  The
    result is a JSON-format character scalar if \code{json} is
    \code{TRUE}; otherwise it is an R list with named components,
    extracted from the JSON return value.
}
\description{
Send a query to the motus API
}
\note{
all queries and return values are logged in the file "motus_query_log.txt"
in the user's home directory.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
