% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusQuickRegTag.R
\name{motusQuickRegTag}
\alias{motusQuickRegTag}
\title{Quickly register a tag with motus}
\usage{
motusQuickRegTag(projectID, mfgID, period, dateBin, codeSet = 4,
  model = NULL, tsStart = NULL, species = NULL, nomFreq = 166.38, ...)
}
\arguments{
\item{projectID:}{integer scalar; motus internal project ID;
Alternatively, a character scalar which is matched against
project names, and if a match is found, that is used, with
prompting.}

\item{mfgID:}{character scalar; typically a small integer; return
only records for tags with this manufacturer ID (usually
printed on the tag).  If a duplicate within a given project and
season, it should have decimal digits indicating the number of
physical marks added to the tag.}

\item{period:}{numeric scalar; approximate repeat interval of tag
transmission, in seconds, to nearest 0.1 s; the precise value
is guessed from the tag database}

\item{dateBin:}{character scalar; quick and dirty approach to
keeping track of tags likely to be active at a given time; this
records "YYYY-Q" where Q is 1, 2, 3, or 4.  Represents the
approximate quarter during which the tag is expected to be
active.  Used in lieu of deployment information when that is
not (yet) available.}

\item{codeSet:}{3 or 4, or their character equivalents.  Default is
4.}

\item{model:}{one of the Lotek model codes.  Default is NULL,
meaning unknown.}

\item{tsStart:}{start of deployment, if known.}

\item{species:}{4-letter code of species, if known}

\item{...:}{additional parameters to motusQuery()}
}
\value{
a 3-element numeric vector; the first element is the motus
    tag ID, and the second element, if not NA, is the motus
    deployment ID.  The third element is the estimated true burst
    interval, in seconds.
}
\description{
This lets you register a tag with minimal fuss, provided you know
the project name or ID, manufacturer's tag ID, and approximate
burst interval (to 0.1 s precision).  You should also know the
dateBin, i.e. the year and quarter the tag was deployed in.
}
\details{
This function uses the official Lotek database to look up tag
parameters.  It also estimates a precise burst interval as the mean
of those burst intervals from previously-registered tags which are
within 0.05 s of the one you specify here.

You will be shown the full set of registration parameters and given
a chance to cancel the process if they look wrong.
}
\note{
: if both \code{tsStart} and \code{species} are NULL, then a
    tag deployment record is \emph{not} generated for the
    newly-registered tag.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
