% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusRegisterProject.R
\name{motusRegisterProject}
\alias{motusRegisterProject}
\title{register a new motus project}
\usage{
motusRegisterProject(projectName, ...)
}
\arguments{
\item{projectName:}{character scalar; human-readable name of project}

\item{...:}{additional parameters to motusQuery()}
}
\description{
register a new motus project
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

