% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusRegisterReceiver.R
\name{motusRegisterReceiver}
\alias{motusRegisterReceiver}
\title{register a receiver with motus (admin only)}
\usage{
motusRegisterReceiver(serno)
}
\arguments{
\item{serno:}{character vector, such as "SG-1234BBBK5678"}
}
\value{
the query result, which will be a list.  If registration is successful,
the list will have an integer item named `deviceID`, which is the motus
ID for the receiver.
}
\description{
register a receiver with motus (admin only)
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
