% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusRegisterTag.R
\name{motusRegisterTag}
\alias{motusRegisterTag}
\title{register a tag with motus}
\usage{
motusRegisterTag(projectID, mfgID, manufacturer = "Lotek", type = "ID",
  codeSet = "Lotek4", offsetFreq, period, periodSD, pulseLen, param1, param2,
  param3, param4, param5, param6, paramType = 1, ts, nomFreq, dateBin, model,
  ...)
}
\arguments{
\item{projectID:}{integer scalar; motus internal project ID}

\item{mfgID:}{character scalar; typically a small integer; return
only records for tags with this manufacturer ID (usually
printed on the tag)}

\item{type:}{character scalar; tag type; "ID" or "beeper"}

\item{codeSet:}{character scalar; manufacturer's codeset label,
when \code{type} is "ID"}

\item{offsetFreq:}{numeric scalar; offset of tag frequency from
nominal, in kHz}

\item{period:}{numeric scalar; repeat interval of tag transmission,
in seconds.}

\item{periodSD:}{numeric scalar; standard deviation of period}

\item{pulseLen:}{numeric scalar; length of pulses emitted by tag,
in milliseconds}

\item{param1:}{numeric scalar; first measured tag parameter; for
Lotek coded ID, this is the first interpulse gap}

\item{param2:}{numeric scalar; second measured tag parameter; for
Lotek coded ID, this is the second interpulse gap}

\item{param3:}{numeric scalar; third measured tag parameter; for
Lotek coded ID, this is the third interpulse gap}

\item{param4:}{numeric scalar; fourth measured tag parameter; for
Lotek coded ID, this is the standard deviation of the first
interpulse gap}

\item{param5:}{numeric scalar; fifth measured tag parameter; for
Lotek coded ID, this is the standard deviation of the second
interpulse gap}

\item{param6:}{numeric scalar; sixth measured tag parameter; for
Lotek coded ID, this is the standard deviation of the third
interpulse gap}

\item{paramType:}{integer scalar; indicates the type of parameters
supplied for this tag; should be a foreign key to a table with
parameter type info.}

\item{ts:}{numeric scalar; registration timestamp; typically
Sys.time(), but for the time at which the registration was
actually made, not just when the API was called.}

\item{nomFreq:}{numeric scalar; nominal tag carrier frequency, in MHz}

\item{dateBin:}{character scalar; quick and dirty approach to
keeping track of tags likely to be active at a given time; this
records "YYYY-Q" where Q is 1, 2, 3, or 4.  Represents the
approximate quarter during which the tag is expected to be
active.  Used in lieu of deployment information when that is
not (yet) available.}

\item{model:}{character scalar; Lotek tag model}

\item{...:}{additional parameters to motusQuery()}
}
\value{
query results.
}
\description{
register a tag with motus
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
