% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusReserveKeys.R
\name{motusReserveKeys}
\alias{motusReserveKeys}
\title{Reserve a block of consecutive primary keys in a motus transfer table.}
\usage{
motusReserveKeys(table, key, n, IDcol, ID)
}
\arguments{
\item{table}{name of the table in which to reserve keys.}

\item{key}{name of the key column in table \code{table}.  This
column must be a INT (or BIGINT) PRIMARY KEY.}

\item{n}{number of consecutive keys to reserve.  If negative, the
keys are reserved in descending order, starting at the smallest
available value.}

\item{IDcol}{name of column in which value \code{ID} is placed.
Placing a receiver-unique value in this column lets us retrieve
the value for the \code{key} column.}

\item{ID}{value unique across receivers that is placed in column
\code{IDcol} to mark it as belonging to this receiver.
Ideally, this is the motusRecvID key for this receiver.}
}
\value{
the first key value in the block.  For positive \code{n}, this
is the lowest key value in the block.  For negative \code{n}, this is
the highest key value. i.e. in both cases, keys are allocated in order
from closest to 0 to farther from 0.
}
\description{
This function returns the first key value in a block of N consecutive
key values for the specified table.  These key values are atomically
reserved, and guaranteed not to collide with key values for any other
receiver.  This permits multiple processes to be pushing data to
motus (via populating the transfer tables) at the same time.
}
\details{
Note: a record with the largest key value for the block, and default
values for other columns, is inserted into the table.
}
\examples{

## return the first key in a block of 200 reserved in the batches table
motusReserveKeys("batches", "batchID", 200, "motusRecvID", 1234)

}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

