% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusSaveSecrets.R
\name{motusSaveSecrets}
\alias{motusSaveSecrets}
\title{save secrets for motus web API in default location}
\usage{
motusSaveSecrets()
}
\description{
If the variable MOTUS_SECRETS is not NULL, its contents are stored
in JSON format in the package file motusSecrets.json, from where
they will be automatically read when the \code{motus} package is
loaded.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
