% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusTempPath.R
\name{motusTempPath}
\alias{motusTempPath}
\title{Create a temporary filename or folder for the motus server.}
\usage{
motusTempPath(isdir = TRUE)
}
\arguments{
\item{isdir}{boolean scalar: if TRUE, the default, a temporary
folder is created; otherwise, a file is created}
}
\value{
the path to the new file or folder
}
\description{
Create a temporary filename or folder for the motus server.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

