% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveDirContents.R
\name{moveDirContents}
\alias{moveDirContents}
\title{Move the contents of a directory to another directory.}
\usage{
moveDirContents(src, dst)
}
\arguments{
\item{src}{character scalar; path to source directory}

\item{dst}{path to target folder}
}
\value{
a boolean vector of the same length as
\code{list.files(src, all.files=TRUE, recursive=FALSE, no..=TRUE, include.dirs=TRUE)}, with TRUE
entries corresponding to files moved successfully.
}
\description{
Move the contents of a directory to another directory.
}
\details{
All files and sub-directories are moved.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

