% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveFiles.R
\name{moveFiles}
\alias{moveFiles}
\title{Move files/folders to a folder}
\usage{
moveFiles(src, dst)
}
\arguments{
\item{src}{character vector; paths to source files / folders}

\item{dst}{character scalar; path to target folder}
}
\value{
a boolean vector of the same length as \code{src}, with
    TRUE entries corresponding to files/folders moved successfully.
}
\description{
Move files/folders to a folder
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

