% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveJob.R
\name{moveJob}
\alias{moveJob}
\title{move a job from one folder to another}
\usage{
moveJob(job, dest)
}
\arguments{
\item{job}{Twig object representing the job.}

\item{dest}{path to new folder where the job's folder will become
a subfolder.}
}
\value{
TRUE on success; FALSE if the job's folder could not be moved
}
\description{
Retain the old path to the job, in case of interruption
between recording the new path and moving the folder.
Also, move all jobs which have this job as an ancestor,
adjusting their \code{$path} items appropriately
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{Copse}}
}

