% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.motusMap.R, R/getMap.R
\name{names.motusMap}
\alias{names.motusMap}
\title{get the key values of a map table in a database}
\usage{
\method{names}{motusMap}(map)

\method{names}{motusMap}(map)
}
\arguments{
\item{map}{object of class "motusMap"}

\item{map}{object of class "motusMap"}
}
\value{
character vector; all keys in the map

character vector; all keys in the map
}
\description{
get the key values of a map table in a database

get the key values of a map table in a database
}

