% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newJobStep.R
\name{newJobStep}
\alias{newJobStep}
\title{create a new job step.}
\usage{
newJobStep(job, type, params)
}
\arguments{
\item{job}{object of class "motusJob"}

\item{type}{name of handler for this job step}

\item{params}{R object representable in JSON}
}
\value{
This function returns a named integer vector of length 1
    with class "motusJobStep".  The value is the job step number,
    and the name is the full path to the new folder.
}
\description{
A new unique jobStep # is generated, and a folder for it is added
to the folder of the parent job. An entry in the server database is
created.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

