% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSubJob.R
\name{newSubJob}
\alias{newSubJob}
\title{create and enqueue a new subjob of an existing job}
\usage{
newSubJob(j, type, ...)
}
\arguments{
\item{j}{existing job}

\item{type}{character scalar of type of subjob}

\item{...}{additional named metadata for this job}
}
\value{
A Twig (see \link{\code{Copse}}) object \code{j} representing the job.
    The path to the job folder is accessible as \code{j$path}.
}
\description{
The subjob gets the existing job as parent, and will be of
the same type.  The subjob's folder is created inside the
parent's folder.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{Copse}}
}

