% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldSitePath.R
\name{oldSitePath}
\alias{oldSitePath}
\title{get the path to an old-style sensorgnome site.}
\usage{
oldSitePath(year, proj, site)
}
\arguments{
\item{year}{integer; year of receiver deployment}

\item{proj}{character scalar; project short name}

\item{site}{character scalar; site short name}
}
\value{
character vector giving path to each site in the old '/SG/' hierarchy.
There is one entry for each row of \code{paste0(year, proj, site)}
}
\description{
Called by \code{\link{server}} for a file or folder added
to the queue.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

