% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagTimeline.R
\name{plotTagTimeline}
\alias{plotTagTimeline}
\title{plot a timeline of all deployed tags}
\usage{
plotTagTimeline(sort = c("projCode", "dateBin", "sort", "nomFreq", "iMfgID"),
  filename = "/sgm/pub/motus_tag_timeline.html")
}
\arguments{
\item{sort}{character vector of field names by which to sort.  Default:
NULL, which means sort by \code{projCode, sort, nomFreq, mfgID} (i.e.
by manufacturer's tag ID within nominal frequency within taxonomic
species order, within project code.}

\item{filename}{filename to save plot; Default: "/sgm/pub/motus_tag_timeline.html"}
}
\value{
a data_frame with these columns:
\itemize{
\item motusID
\item projectID
\item projCode
\item fullID
\item mfgID
\item nomFreq
\item period
\item speciesID
\item speciesName
...
}
}
\description{
This generates an html file with preformatted text which shows
a tag timeline when viewed in a fixed-width font.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
