% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postInstall.R
\name{postInstall}
\alias{postInstall}
\title{code to run after installing a new version of this package.}
\usage{
postInstall()
}
\description{
After re-installing the motusServer package, this function is executed
to fix ownership, permissions, and links outside the R package tree.
}
\details{
If you install the motusServer R package using this command line:

sudo su -c 'cd /home/sg/src/motusServer; /usr/local/bin/rpack -g .' sg

then this function will be run by that script.  Otherwise, you can
do so manually from the command line like so:

Rscript -e 'motusServer:::postInstall()'
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
