% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{project_ambiguities_for_tag_project}
\alias{project_ambiguities_for_tag_project}
\title{get project ambiguity groups for a given project}
\usage{
project_ambiguities_for_tag_project(env)
}
\arguments{
\item{projectID}{integer scalar project ID}
}
\value{
a list with these vector items:
\itemize{
   \item ambigProjectID; negative integer project ambiguity ID
   \item projectID1; positive integer motus project ID
   \item projectID2; positive integer motus project ID
   \item projectID3; positive integer motus project ID or null
   \item projectID4; positive integer motus project ID or null
   \item projectID5; positive integer motus project ID or null
   \item projectID6; positive integer motus project ID or null
}
}
\description{
get project ambiguity groups for a given project
}
