% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{pulse_counts_for_receiver}
\alias{pulse_counts_for_receiver}
\title{get pulse counts from a batch}
\usage{
pulse_counts_for_receiver(env)
}
\arguments{
\item{batchID}{integer batchID}

\item{ant}{integer}

\item{hourBin}{numeric hourBin of latest pulseCounts already obtained

The pair (ant, hourBin) is for the latest record already obtained.
For each \code{batchID}, records are returned sorted by
\code{hourBin} within \code{ant}.  For the first call with each \code{batchID},
the caller should specify \code{hourBin=0}, in which case \code{ant} is ignored.}
}
\value{
a data frame with the same schema as the pulseCounts table, but
    JSON-encoded as a list of columns
}
\description{
get pulse counts from a batch
}
