% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purgeFromMotus.R
\name{purgeFromMotus}
\alias{purgeFromMotus}
\title{Delete all data for this receiver from ths motus transfer tables.}
\usage{
purgeFromMotus(src)
}
\arguments{
\item{src}{dplyr src_sqlite to receiver database}
}
\value{
no return value.
}
\description{
Deletes all records in all tables of the motus transfer database that
relate to this receiver. This should only be done very rarely.
e.g. during development/testing.  Normally, if a batch needs to be
re-run for some reason, the formal batchDelete mechanism should be
used, so that an audit trail is left.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

