% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queueJob.R
\name{queueJob}
\alias{queueJob}
\title{enqueue a job}
\usage{
queueJob(j)
}
\arguments{
\item{j}{the job to add; e.g. as generated by \link{\code{newJob}}}
}
\value{
no return value
}
\description{
Add a job to the global MOTUS_QUEUE.
}
\note{
If the job is already in the queue, this function does nothing.
If the job is marked "done", it is not added to the queue.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

