% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queueKnownFiles.R
\name{queueKnownFiles}
\alias{queueKnownFiles}
\title{Queue any files of known type from a folder.}
\usage{
queueKnownFiles(dir = character(0), files = character(0))
}
\arguments{
\item{dir}{character vector of directories; default \code{character(0)}}

\item{files}{character vector of full paths to files; default \code{character(0)}}
}
\value{
an integer vector with two elements: the number of
    attachments handled and the total number of message parts.
}
\description{
The email must already have been unpacked into the specified
directory.  Files from these known types are enqueued:
}
\details{
\itemize{
\item .DTA files from a lotek receiver
\item .txt.gz  compressed files from an SG
\item .txt  uncompressed files from an SG
\item .zip compressed archive holding any of the above types
\item .7z ...
\item .rar ...
}

Any remaining files are relegated to manual intervention.
}
\seealso{
\link{\code{server}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
