% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queueUnfinishedSubjobs.R
\name{queueUnfinishedSubjobs}
\alias{queueUnfinishedSubjobs}
\title{Load unfinished jobs from server database and enqueue them.}
\usage{
queueUnfinishedSubjobs(topJob)
}
\arguments{
\item{topJob}{integer; the topjob for which unfinished subjobs will be queued}
}
\value{
the number of subjobs enqueued in the global \code{MOTUS_QUEUE}.
}
\description{
Load unfinished jobs from server database and enqueue them.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

