% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realFile.R
\name{realFile}
\alias{realFile}
\title{follow any symlinks, returning the path to a normal file
or folder}
\usage{
realFile(path)
}
\arguments{
\item{paths}{character vector of paths}
}
\value{
character value of the same length as \code{path}, with any
    symlinks followed so that each item returned is a path to a
    normal file or folder.  For any item in \code{path} which
    doesn't exist, or for which the last symlink target doesn't exist,
    or for which there is a symlink loop, returns \code{NA}.
}
\description{
Given a path, there are some file operations where we want to be
sure we're dealing with the real file, and not a symlink to it.
For a vector of paths, this function follows any symlinks along
each path, returning the ultimate normal file or directory as a
target.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
