% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{receivers_for_project}
\alias{receivers_for_project}
\title{get receivers for a project}
\usage{
receivers_for_project(env)
}
\arguments{
\item{projectID;}{integer scalar project ID}
}
\value{
a list with these vector items:
   \itemize{
      \item projectID; integer ID of project that deployed the receiver
      \item serno; character serial number, e.g. "SG-1214BBBK3999", "Lotek-8681"
      \item receiverType; character "SENSORGNOME" or "LOTEK"
      \item deviceID; integer device ID (internal to motus)
      \item status; character deployment status
      \item name; character; typically a site name
      \item fixtureType; character; what is the receiver mounted on?
      \item latitude; numeric (initial) location, degrees North
      \item longitude; numeric (initial) location, degrees East
      \item elevation; numeric (initial) location, metres ASL
      \item isMobile; integer non-zero means a mobile deployment
      \item tsStart; numeric; timestamp of deployment start
      \item tsEnd; numeric; timestamp of deployment end, or NA if ongoing
   }
}
\description{
get receivers for a project
}
