% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordEvent.R
\name{recordEvent}
\alias{recordEvent}
\title{record a processing event in the history table of the motus transfer database}
\usage{
recordEvent(event = c("MERGE", "FINDTAGS", "CLEAN", "COMPAREOLDNEW", "PLOT"),
  origin, serno = "", exitCode = 0, errorMsg = "", URLs = "",
  URLlabels = "")
}
\arguments{
\item{event}{- short character scalar giving type of event; c("CLEAR", "MERGE", "FIND")}

\item{origin}{- character scalar describing source of event; e.g. "data email from stuart@bsc-eoc.org", "user command on sensorgnome.org"}

\item{serno}{- character scalar giving full serial number of device involved, if any. e.g. "SG-3214BBBK1512"}

\item{exitCode}{- integer scalar giving exit code of processing; 0 means success; otherwise an error number. Default: 0.}

\item{errorMsg}{- character scalar giving error message, if any.  Default: NULL}

\item{URLs}{- character vector giving URLs of any output files posted to website}

\item{URLlabels}{- character vector giving descriptive label (e.g. filename) for each URL}
}
\value{
no return value.  Side effect: a record is added to the \code{history} table in the motus transfer database
}
\description{
Recorded events are used to generate history pages on the wiki.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

