% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshMotusMetaDBCache.R
\name{refreshMotusMetaDBCache}
\alias{refreshMotusMetaDBCache}
\title{Refresh our cache of the motus database of metadata for tags,
receivers, projects, and species.}
\usage{
refreshMotusMetaDBCache()
}
\value{
the number of tables updated.
}
\description{
This refresh is wrapped in a transaction.  Before the transaction
is committed, we commit the git repo version of the DB, and
grab that commit hash.
}
\details{
For each table X, if the appropriate query fails, we leave X as-is.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
