% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexPieces.R
\name{regexPieces}
\alias{regexPieces}
\title{Extract named items from strings according to a regular expression.}
\usage{
regexPieces(rx, s)
}
\arguments{
\item{rx}{Perl-type regular expression with named fields, as
described in \code{?regex}}

\item{s}{character vector.}
}
\value{
a list of character vectors.  Each vector corresponds to an
    item of \code{s}.  The vector items are substrings of the input
    item corresponding to named capture groups in \code{rx}, in the
    order they appear in the input item.  The vector names are
    those of the capture group.  The same name might appear more
    than once in the vector.  If an item in the input does not
    match any capture of \code{rx}, then its corresponding vector
    in the output is empty.
}
\description{
FIXME (eventually): when the stringi package regex code can handle
named subexpressions, use \code{stri_extract_all_regex(...,
simplify=TRUE)}
}
\examples{

rx = "(?<ayes>a+)|(?<bees>b+)|(?<sees>c+)"
s = c("anybbody", "something", "aebbfcbggaa")
regexPieces(rx, s)
## returns:
##  [[1]]
##  ayes bees
##   "a" "bb"
##
##  [[2]]
##  named character(0)
##
##  [[3]]
##  ayes bees sees bees ayes
##   "a" "bb"  "c"  "b" "aa"

}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

