% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerProducts.R
\name{registerProducts}
\alias{registerProducts}
\title{register product files for a receiver}
\usage{
registerProducts(j, path, serno = j$serno, projectID, isTesting = FALSE)
}
\arguments{
\item{j}{the job}

\item{path}{character vector; locations of the product file}

\item{serno}{serial number of receiver these products' data come from;
default: j$serno}

\item{projectID}{ID of project that owns the product(s)}

\item{isTesting}{are these test products?  Default: FALSE}
}
\value{
TRUE
}
\description{
register product files for a receiver
}
\details{
Records the products in the job database and creates a symlink
to it from the appropriate download folder.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
