% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparentJob.R
\name{reparentJob}
\alias{reparentJob}
\title{move a subjob to a new parent job}
\usage{
reparentJob(j, p)
}
\arguments{
\item{j}{Twig object representing the subjob.}

\item{p}{Twig object representing the new parent.  If}
}
\value{
TRUE on success; FALSE if the job's folder could not be moved

Note: this function can not reparent a top job.
}
\description{
Also, if the job has a folder, then the job's folder is moved from
its current location to the the folder of the new parent.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

