% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retryJob.R
\name{retryJob}
\alias{retryJob}
\title{Retry a job, after some external error condition has been fixed}
\usage{
retryJob(j, msg = "external error corrected")
}
\arguments{
\item{j}{the job, as an integer scalar job number. This can be a
top-level job, or one of its sub-jobs; in the latter case, the
top-level job is used anyway.}

\item{msg}{a message explaining how the error condition was fixed.}
}
\value{
TRUE if the job was found and resubmitted to queue 0.
FALSE otherwise.
}
\description{
Jobs often end in an error that requires external intervention.
After such intervention has taken place, this function can be
called to re-run those subjobs which either had an error, or
which were run after a subjob with an error.
}
\details{
All subjobs of \code{j} with errors will be re-run.
In addition, any subjob of \code{j} which ran after a subjob
with errors will also be re-run.

To do this, the function:
\itemize{
\item sets \code{sj$done} to 0 for any subjob \code{sj} where
\code{sj$done < 0} or for any subjob last run after such a subjob
\item sets \code{j$path} to \code{MOTUS_PATH$QUEUEx} where x is the value of the job's
queue.
\item moves the job folder from its current location to \code{MOTUS_PATH$QUEUEx}
}

One of the process servers watching queue 0 will then claim the job
and re-run those subjobs touched above, as well as any which had not
been run.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{processServer}}
}

