% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copse.R
\name{Reval}
\alias{Reval}
\title{evaluate portions of an expression enclosed in \code{R( )},
returning the reduced expression.}
\usage{
Reval(e)
}
\arguments{
\item{e}{an expression.}
}
\description{
This allows query expressions to include portions to be
evaluated in R before converting the remaining expression into
an SQLite:json1 query
}

