% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runningCount.R
\name{runningCount}
\alias{runningCount}
\title{running count of occurences for each unique value in a vector}
\usage{
runningCount(x)
}
\arguments{
\item{x:}{integer vector}
}
\value{
integer vector \code{rv}; \code{rv[i] =
    sum(x[1:i] == x[i])}
}
\description{
running count of occurences for each unique value in a vector
}
\details{
For an integer vector, returns a vector of equal length giving the
    number of times the i'th element occurs in the first i slots.
}
\note{
: despite the O(log(length(x))) call to \code{order()},
    this works faster up to at least length(x) == 1E8 than a similar
    version that first converts to a factor and then orders using
    method="radix"
}
\examples{

runID = ceiling(10 * runif(100))
posInRun = runningCount(runID)
for (i in seq(along=runID)) if (posInRun[i] != sum(runID[1:i] == runID[i])) stop("runningCount failed!")
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

