% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{runs_for_receiver}
\alias{runs_for_receiver}
\title{get all runs from a batch for a receiver}
\usage{
runs_for_receiver(env)
}
\arguments{
\item{batchID}{integer batchID}

\item{runID}{integer ID of largest run already obtained}
}
\value{
a data frame with the same schema as the runs table, but JSON-encoded as a list of columns
}
\description{
get all runs from a batch for a receiver
}
