% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{runs_for_tag_project}
\alias{runs_for_tag_project}
\title{get runs by tag project from a batch}
\usage{
runs_for_tag_project(env)
}
\arguments{
\item{projectID}{integer project ID}

\item{batchID}{integer batchID}

\item{runID}{integer ID of largest run already obtained}
}
\value{
a data frame with the same schema as the runs table, but JSON-encoded as a list of columns
}
\description{
get runs by tag project from a batch
}
