% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSGbatch.R
\name{runSGBatch}
\alias{runSGBatch}
\title{Process a batch of raw sensorgnome receiver files.  The output
goes into the mysql "motus" database.}
\usage{
runSGBatch(files, haveTagDeployments)
}
\arguments{
\item{files}{character vector of full paths to all files in batch}

\item{haveTagDeployments}{boolean scalar: do we have all deployment
data for tags?  If so, the list of tags to be searched for is
obtained from motus based on deployment dates, otherwise, it is
based on registration dates.}
}
\description{
A batch of files are all from the same receiver and boot session.
This code queries the motus-wts API for metadata:
}
\details{
- receiver ID (based on serial number, and eventually MAC address)

- tag database (based on dates and eventually location); list of
  all tags to be sought over the period of time covered by the
  batch of files.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

