% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_arg.R
\name{safe_arg}
\alias{safe_arg}
\title{Extract an argument from a list ensuring SQL-safety}
\usage{
safe_arg(x, name, type, scalar = TRUE)
}
\arguments{
\item{x}{list object with names}

\item{name}{unquoted object name}

\item{type}{unquoted type name, one of "character", "integer", "logical", "numeric", "list"
or abbreviation thereof.}

\item{scalar}{if TRUE (the default) return a vector of length 1,
dropping any additional items. Otherwise, return a vector with the
same length as the item from \code{x}.  If \code{type=="list"}, the
vector is of mode "list".}
}
\value{
If there is no item with the given name in \code{x},
then return NULL.
}
\description{
If the list contains an item of the given name, the return value
will be that item coerced to the given type. For character strings,
single quotes will be doubled so that the object is safe for use in
SQL queries, provided that strings are enclosed in single quotes in
the query.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
