% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeFileRename.R
\name{safeFileRename}
\alias{safeFileRename}
\title{rename files, even across filesystems}
\usage{
safeFileRename(from, to)
}
\arguments{
\item{from}{character vector giving filename or path}

\item{to}{character vector giving filename or path}
}
\value{
logical vector,\code{rv}, that is TRUE for each successful rename, FALSE elsewhere

Upon returning:
\itemize{
   \item the files `from[rv]` no longer exist, and their contents are now stored in files `to[rv]`.
   \item the files `from[! rv]` and `to[! rv]` are preserved, if they already existed, and
are not created if they didn't already exist.
}
i.e. unsuccesful renaming leaves source and destination items as they were.
}
\description{
drop-in replacement for file.rename that correctly handles
renames across filesystems on linux
- copies the file from filesystem A to filesystem B
- only if the copy succeeded, delete the file from filesystem B
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
