% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeSQL.R
\name{safeSQL}
\alias{safeSQL}
\title{Return a function that safely performs sql queries on a connection.}
\usage{
safeSQL(con, busyTimeout = 300)
}
\arguments{
\item{con}{RSQLite connection to database, as returned by
dbConnect(SQLite(), ...), or character scalar giving path
to SQLite database, or MySQLConnection.}

\item{busyTimeout}{how many total seconds to wait while retrying a
locked database.  Default: 300 (5 minutes).  Uses \code{pragma busy_timeout}
to allow for inter-process DB locking.  Only implemented for
SQLite connections.}
}
\description{
This uses dbGetPreparedQuery (for RSQLite) or dbEscapeStrings
(for MySQL).  It should prevent e.g. SQL injection attacks.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

