% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeURL.R
\name{sanitizeURL}
\alias{sanitizeURL}
\title{rewrite a URL to remove any authorization credentials.}
\usage{
sanitizeURL(url, method = "generic")
}
\arguments{
\item{url}{character scalar; location of downloadable resource}

\item{method}{character scalar; method for downloadable resource,
as determined from \code{dataTransferRegex}.  Defaults to "generic",
which does a generic sanitization.}
}
\value{
\code{url} with sufficient portions removed to prevent unauthorized use.
}
\description{
The resulting URL will still look enough like the original to allow
a user with access to the original to match it to the sanitized version.
}
\details{
When users send data files, we'd like to have a public record of
downloading and processing, but URLs can contain implicit or explicit
credentials which we don't want to make public.

This function applies an appropriate method to the given URL to ensure this.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
