% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{sendError}
\alias{sendError}
\title{send an error as the reply}
\usage{
sendError(res, error)
}
\arguments{
\item{res}{Rook::Response object}

\item{error}{character vector with error message(s)}
}
\value{
no return value
}
\description{
send an error as the reply
}

