% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{sendHeader}
\alias{sendHeader}
\title{send the header for a reply}
\usage{
sendHeader(res)
}
\arguments{
\item{res}{Rook::Response object}
}
\value{
no return value
}
\description{
send the header for a reply
}
\details{
the reply will always be a gzip-compressed JSON-encoded value.
We also disable caching.
}
