% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusServer2.R
\name{serno_collision_rules}
\alias{serno_collision_rules}
\title{get, create, or delete rules for resolving receiver serial number collisions}
\usage{
serno_collision_rules(env)
}
\description{
get, create, or delete rules for resolving receiver serial number collisions
}
