% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{wait for new data files to arrive, then process them}
\usage{
server(handlers, logFun)
}
\arguments{
\item{handlers}{list of function to be called when a new file or
folder is added to \code{/sgm/incoming}.  Each function must
accept a single chararacter parameter, which will be a path to
a temporary directory containing the new file(s).  Each
function must return TRUE if all files were processed
successfully, or FALSE otherwise.  Returning FALSE will prevent
the temporary directory and its files from being deleted.
Handlers are called once for each file or folder already in the
watch directory, and then once for each file or folder created
in, moved into, or linked to from that directory.  Handlers are
permitted to copy, move, or delete files, but must not modify
them in-place.}

\item{logFun}{a function to which log messages are
passed.  Defaults to a function which uses cat()s its arguments to
stderr.  \code{logFun} should have the parameter list "..."}
}
\value{
This function does not return.
}
\description{
This function watches the /sgm/incoming directory and when a file
or folder is written to it, carries out the specified action.
Files or folders already in the watched directory are processed
before any new ones, on the assumption that a previous call to this
function was interrupted.
}
\details{
The function does not return; it is meant for use in an R script
run in the background.
}
\note{
If a directory is added by creating a symlink to it in
    \code{/sgm/incoming}, ownership of files in the directory
    remains with the files' creator, and this function will not
    delete them.  Otherwise, ownership is assumed by this function,
    and after all handlers have been called, files are deleted.
    Each handler is passed a directory populated with hardlinks to
    the original files.  That way, the handler can move or delete
    the files without consequence to either other handlers or the
    caller.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

