% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverCommon.R
\name{serverCommon}
\alias{serverCommon}
\title{common code required by all XXXServer() functions}
\usage{
serverCommon()
}
\value{
TRUE
}
\description{
common code required by all XXXServer() functions
}
\note{
side effects are to load libraries, open databases, assign to global variables
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
