% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverError.R
\name{serverError}
\alias{serverError}
\title{Handle an error for the motus server.}
\usage{
serverError()
}
\value{
character scalar summary error message
}
\description{
This function is installed as the top-level error handler by
\link{\code{server}}
}
\details{
It records a stack dump to /sgm/errors and returns a summary error
message.
}
\note{
based on R's dump.frames()
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

