% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgEnsureDBTables.R
\name{sgEnsureDBTables}
\alias{sgEnsureDBTables}
\title{make sure a receiver database has the required tables; also, load
custom SQLite extensions on this DB connection.}
\usage{
sgEnsureDBTables(src, recreate = c())
}
\arguments{
\item{src}{dplyr sqlite src, as returned by \code{dplyr::src_sqlite()}}

\item{recreate}{vector of table names which should be dropped then re-created,
losing any existing data.  Defaults to empty vector, meaning no tables
are recreate.  As a special case, TRUE causes all tables to be dropped
then recreated.}
}
\value{
returns NULL (silently); fails on any error
}
\description{
make sure a receiver database has the required tables; also, load
custom SQLite extensions on this DB connection.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
