% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgFilenameRegex.R
\docType{data}
\name{sgFilenameRegex}
\alias{sgFilenameRegex}
\title{Regular expression for parsing data filenames from a sensorgnome receiver.}
\format{An object of class \code{character} of length 1.}
\usage{
sgFilenameRegex
}
\description{
Regular expression for parsing data filenames from a sensorgnome receiver.
}
\keyword{datasets}
