% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgMergeFiles.R
\name{sgMergeFiles}
\alias{sgMergeFiles}
\title{Merge a batch of raw SG files with their receiver database(s).}
\usage{
sgMergeFiles(files, j, dbdir = MOTUS_PATH$RECV)
}
\arguments{
\item{files}{either a character vector of full paths to files, or
the full path to a directory, which will be searched
recursively for raw sensorgnome data files.}

\item{j}{job, whose ID will be recorded with records of new / changed files.}

\item{dbdir}{path to folder with existing receiver databases
Default: \code{MOTUS_PATH$RECV}}
}
\value{
a list with two items:
\itemize{
\item info- a data_frame reporting the details of each file, with these columns:
\itemize{
\item name - full path to filename
\item use - TRUE iff data from this file will be incorporated in this run (i.e. the file is new, or has
more content than its previous version)
\item new  - TRUE iff a file of this name was not yet in database
\item done  - TRUE iff this is a compressed file and we have its full contents
\item corrupt      - TRUE the file was compressed but corrupted
\item small - TRUE iff the file contents are shorter than existing contents for that file in the receiver DB
\item partial - TRUE iff this is a partial compressed file for which an uncompressed version is in the same batch
\item serno - serial number of the receiver, parsed from this filename
\item monoBN - monotonic boot session, parsed from filename
\item ts - starting timestamp, parsed from filename
}
\item resumable - a logical vector indicating the tag finder can be run with --resume for each
subset of files from a distinct (serno, monoBN) pair.  The vector has names consisting of
\code{paste(serno, monoBN)}.
}

Returns NULL if no valid sensorgnome data files were found.
}
\description{
Determines which files are redundant, new, or partially new.  Does
*not* run the tag finder.
}
\details{
Any new content from files is merged into the files and fileContents
tables of the receiver DBs.

Files are disposed of like so:
\itemize{
\item any files which are symlinks are deleted
after merging their target's contents; i.e. the symlink is deleted,
not the file it points to.
\item files whose path has \code{MOTUS_PATH$FILE_REPO} as a prefix are left as-is
\item files whose path does not have \code{MOTUS_PATH$FILE_REPO} as a prefix, and which
either are new or have new content are moved to \code{MOTUS_PATH$FILE_REPO/serno/YYYY-MM-DD}
\item remaining files are moved to \code{MOTUS_PATH$TRASH}
}
}
\seealso{
\code{sgRunNewFiles}, which calls this function and then calls \code{sgFindTags}
as appropriate.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
