% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgRawDataRegex.R
\docType{data}
\name{sgRawDataRegex}
\alias{sgRawDataRegex}
\title{Regular expression for parsing line from raw data file.}
\format{An object of class \code{character} of length 1.}
\usage{
sgRawDataRegex
}
\description{
Regular expression for parsing line from raw data file.
}
\keyword{datasets}

