% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgRecvSrc.R
\name{sgRecvSrc}
\alias{sgRecvSrc}
\title{Get the src_sqlite for a receiver database given its serial number.}
\usage{
sgRecvSrc(serno, dbdir = MOTUS_PATH$RECV)
}
\arguments{
\item{serno}{raw sensorgnome serial number (i.e. without "SG-")}

\item{dbdir}{path to folder with existing receiver databases
Default: \code{MOTUS_PATH$RECV}}
}
\value{
a src_sqlite for the receiver; if the receiver is new, this database
will be empty, but have the correct schema.
}
\description{
receiver database files are stored in a single directory, and
have names like "SG-XXXXBBBKYYYY.motus".
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

