% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgRunNewFiles.R
\name{sgRunNewFiles}
\alias{sgRunNewFiles}
\title{Process a new batch of files from a receiver.}
\usage{
sgRunNewFiles(files, dbdir = MOTUS_PATH$RECV, ...)
}
\arguments{
\item{files}{either a character vector of full paths to files, or
the full path to a directory, which will be searched
recursively for raw sensorgnome data files.}

\item{dbdir}{path to folder with existing receiver databases
Default: \code{MOTUS_PATH$RECV}}

\item{...}{additional parameters to the tag finder; see \link{\code{sgFindTags}}}
}
\value{
the value returned from \code{sgMergeFiles(files, dbdir)},
    desribing how files were processed.
}
\description{
Merge the files into the receiver database, and run the tag
finder on each boot session for which new input was provided,
either in new files, or longer versions of existing files.
}
\details{
If all usable files for a boot session are more recent than files we
already have, the tag finder is run in "resume" mode, so that only
the new files are processed.  This is the usual situation,
especially for online SGs, and avoids the growth in processing time
that would occur if boot sessions were re-run from scratch with
each incremental file upload.

Otherwise, if any new files for a boot session predate existing
ones, the entire boot session is re-run.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{sgMergeFiles} and \code{sgFindTags}, which this function calls.
}

