% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgRunStream.R
\name{sgRunStream}
\alias{sgRunStream}
\title{Run a function on one or more streams of raw data.}
\usage{
sgRunStream(src, f, mbn = NULL, user = NULL)
}
\arguments{
\item{src}{dplyr src_sqlite to receiver database}

\item{f}{function with this signature:
\enumerate{
\item mbn integer monotonic boot number
\item fts numeric file timestamp; the timestamp encoded in the file name
\item cno chunk number, as follows:
\enumerate{
\item negative; chunk and timestamp invalid; called with this value only once, before first chunk; value is number of chunks in stream
\item positive; chunk is valid; \code{cno} is index of chunk, starting at 1
\item 0L chunk and timestamp invalid; called with this value only once, after last chunk; return value for this call becomes return value of \code{sgRunStream}
}

\item dat character scalar; the file contents, with embedded newlines

\item user; the user object passed to \code{sgRunStream}.  This can
be used to maintain state information across calls to \code{f}.
Typically, this would be an environment or reference class object.
}}

\item{mbn}{integer vector of monotonic boot numbers; this is the
monoBN field from the \code{files} table in the receiver's
sqlite database.  Defaults to NULL, which means run all streams
in order.}

\item{user}{arbitrary object passed to \code{f} as \code{user}
parameter on each call.  Defaults to NULL.}
}
\value{
a list with one element per stream; the element is the
    valued returned by \code{f} when called with \code{state=2L};
    i.e. after \code{f} has been called with every chunk for the
    given stream.  All previous return values of \code{f} are
    discarded.  Any side-effects or accumulation of return value
    must be performed by \code{f}.
}
\description{
A stream of raw data consists of all records from raw files with
the same monoBN (montonic boot number), ordered by increasing date.
This function lets you run a function on one or more streams,
specified by monoBN, or all streams.  Each stream is broken into
chunks corresponding to the original files.  Each chunk is passed
as a character scalar which will usually consist of multiple lines,
separated by \code{'\n'}.
}
\note{
any notes
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
