% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgStreamToFile.R
\name{sgStreamToFile}
\alias{sgStreamToFile}
\title{Send a stream of raw data to a file}
\usage{
sgStreamToFile(src, f, mbn = NULL, append = FALSE)
}
\arguments{
\item{src}{dplyr src_sqlite to receiver database}

\item{f}{output filename}

\item{mbn}{integer vector of monotonic boot numbers; this is the
monoBN field from the \code{files} table in the receiver's
sqlite database.  Defaults to NULL, which means run all streams
in order.}

\item{append}{logical scalar; if TRUE, append data to existing file
contents.  Defaults to FALSE.}
}
\value{
invisible NULL
}
\description{
A stream of raw data consists of all records from raw files with
the same monoBN (montonic boot number), ordered by increasing date.
This function lets you write one or more streams,
specified by monoBN, or all streams, to a file.  Typically,
this file is a fifo with a reader, such as the tag finder.
}
\note{
any notes
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
