% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgTimeAtBoot.R
\name{sgTimeAtBoot}
\alias{sgTimeAtBoot}
\title{Get start times for boot sessions for a sensorgnome.}
\usage{
sgTimeAtBoot(serno, monoBN)
}
\arguments{
\item{serno}{character scalar; serial number of sensorgnome, e.g. "SG-5133BBBK2972"}

\item{monoBN}{integer vector; boot session(s) for which boot times are desired.}
}
\value{
real vector of timestamps, one per item of \code{monoBN}.  NA for any
slot where the boot time could not be estimated; e.g. because the receiver
did not get a GPS fix.
}
\description{
Get start times for boot sessions for a sensorgnome.
}
\details{
the start time is typically \code{fixedBy + tsLow} for records in the
receiver's timeFixes table.  This needs to be made explicit; see:
\link{https://github.com/jbrzusto/find_tags/issues/12}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

