% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{size_of_update_for_receiver}
\alias{size_of_update_for_receiver}
\title{get count of update items for a receiver}
\usage{
size_of_update_for_receiver(env)
}
\arguments{
\item{deviceID}{integer motus device ID}

\item{batchID}{integer batchID; only batches with larger batchID are considered}
}
\value{
a list with these items:
\itemize{
\item numBatches
\item numRuns
\item numHits
\item numGPS
}
}
\description{
get count of update items for a receiver
}
