% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleExportPulses.R
\name{sqliteCloneIntoDB}
\alias{sqliteCloneIntoDB}
\title{clone an sqlite table schema (including indexes) from the main DB into an attached DB}
\usage{
sqliteCloneIntoDB(sql, tableName, dbName)
}
\arguments{
\item{sql}{safeSQL object to sqlite database}

\item{tableName}{character; name of table in main database which should be cloned
into attached database}

\item{dbName}{name of attached DB; i.e. the 'D' from a statement
like `attach database 'blam.sqlite' as D`.  The database must already have
been attached by the caller.}
}
\description{
clone an sqlite table schema (including indexes) from the main DB into an attached DB
}
\details{
creates the empty table with the same key and with any associated indexes,
but only if a table (and/or indexes) of the same name don't already exist in the
attached db.
}
