% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusServer2.R
\name{status_api_info}
\alias{status_api_info}
\title{return information about the status api}
\usage{
status_api_info(env)
}
\value{
a list with these items:
   \itemize{
      \item maxRows; integer maximum number of rows returned by other API calls
   }
}
\description{
return information about the status api
}
