% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusServer.R
\name{statusServer}
\alias{statusServer}
\title{reply to http requests for information on the processing queue}
\usage{
statusServer(port, tracing = FALSE)
}
\arguments{
\item{port}{integer; local port on which to listen for requests}

\item{tracing}{logical; if TRUE, run interactively, allowing local user
to enter commands.}
}
\value{
does not return; meant to be run as a server.
}
\description{
reply to http requests for information on the processing queue
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
