% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z[[.motusMap.R
\name{[[.motusMap}
\alias{[[.motusMap}
\title{Return the value of an item from a motusMap.}
\usage{
\method{[[}{motusMap}(map, key)
}
\arguments{
\item{map}{object of class "motusMap"}

\item{key}{character scalar; name of item in map}
}
\value{
character scalar; value associated with \code{key}.
}
\description{
Return the value of an item from a motusMap.
}

