% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMap.R
\name{[[<-.motusMap}
\alias{[[<-.motusMap}
\title{set an element in a motus map}
\usage{
\method{[[}{motusMap}(map, key) <- value
}
\arguments{
\item{map}{object of class "motusMap", as created by \code{motusMap}}

\item{key}{key of item, as character scalar}

\item{val}{character string value to associate with \code{key}}
}
\description{
Set the value associated with a key in the given map.
}
