% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copse.R
\name{[<-.Copse}
\alias{[<-.Copse}
\title{To support syntactic sugar like Jobs[! done && type=="email"]$done = 1,
create idempotent methods:}
\usage{
\method{[}{Copse}(C, expr, sort) <- value
}
\description{
To support syntactic sugar like Jobs[! done && type=="email"]$done = 1,
create idempotent methods:
}
