% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{tags_for_ambiguities}
\alias{tags_for_ambiguities}
\title{get motus tagIDs for ambiguity IDs}
\usage{
tags_for_ambiguities(env)
}
\arguments{
\item{ambigIDs}{integer vector of ambiguity IDs, which are all negative}
}
\value{
a list with these vector items:
\itemize{
   \item ambigID; negative integer tag ambiguity ID
   \item motusTagID1; positive integer motus tag ID
   \item motusTagID2; positive integer motus tag ID
   \item motusTagID3; positive integer motus tag ID or null
   \item motusTagID4; positive integer motus tag ID or null
   \item motusTagID5; positive integer motus tag ID or null
   \item motusTagID6; positive integer motus tag ID or null
}
}
\description{
get motus tagIDs for ambiguity IDs
}

