% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFileContents.R
\name{textFileContents}
\alias{textFileContents}
\title{return the contents of a text file as a character scalar, or the
empty string if the file doesn't exist.}
\usage{
textFileContents(f)
}
\arguments{
\item{f}{full path to the file}
}
\value{
character scalar with the file contents, or empty if the file is empty
or doesn't exist.
}
\description{
return the contents of a text file as a character scalar, or the
empty string if the file doesn't exist.
}
\note{
mainly for reading a file which is not known to exist, e.g. a file whose name
is passed as a parameter to a sub-shell, but where there's no guarantee the subshell
actually wrote to it.

Contents only up to the first embedded NUL are returned.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

