% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toInbox.R
\name{toInbox}
\alias{toInbox}
\title{Move a message to the inbox, from where the emailServer will process it.}
\usage{
toInbox(path)
}
\arguments{
\item{path}{the full path to the message file}
}
\value{
TRUE
}
\description{
Move a message to the inbox, from where the emailServer will process it.
}
\seealso{
\code{\link{emailServer}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
