% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toIncoming.R
\name{toIncoming}
\alias{toIncoming}
\title{Add a file or folder to the external server queue.}
\usage{
toIncoming(path)
}
\arguments{
\item{path}{the full path to the file or directory.

After this call, the file or directory will no longer exist at the
same location, but will be moved into the server's incoming
directory.}
}
\value{
TRUE on success; FALSE otherwise
}
\description{
Add a file or folder to the external server queue.
}
\examples{
## Not run:
toIncoming("/tmp/mytmpurlfile")

}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{server}}
}

