% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topJob.R
\name{topJob}
\alias{topJob}
\title{return the top-level job for which the given job is a subjob}
\usage{
topJob(j)
}
\arguments{
\item{j}{the job}
}
\value{
the toplevel ancestor of \code{j}; this might be
\code{j} itself.
}
\description{
return the top-level job for which the given job is a subjob
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
