% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTrash.R
\name{toTrash}
\alias{toTrash}
\title{send files to the trash folder for eventual deletion}
\usage{
toTrash(f)
}
\arguments{
\item{f}{character vector of full paths to files}
}
\value{
TRUE
}
\description{
We keep all "dispoable" files in one folder.  They are retained there
until we can verify they are no longer needed.  Perhaps we'll reap
them periodically after X weeks.  The trash file is \code{MOTUS_PATH$TRASH}.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
