% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimCommonPath.R
\name{trimCommonPath}
\alias{trimCommonPath}
\title{Remove any common leading path from a set of paths.}
\usage{
trimCommonPath(files)
}
\arguments{
\item{files}{character vector of full paths to files}
}
\value{
\code{files}, with any leading common path removed.
}
\description{
Remove any common leading path from a set of paths.
}
\examples{
trimCommonPath(c("/A/B/C", "/A/B/D/E", "/A/B/F")) == c("C", "D/E", "F")

}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

