% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{TS}
\alias{TS}
\title{bless a numeric vector with "timestamp" class}
\usage{
TS(x)
}
\arguments{
\item{x}{numeric vector; should be seconds since the unix epoch (1 Jan 1970 GMT).}
}
\value{
structure(x, class=c("POSIXt", "POSIXct", class(x)))
}
\description{
adds \code{c("POSIXt", "POSIXct")} to the beginning of the class
attribute of a numeric vector so that operations such as arithmetic
and printing treat it as a timestamp.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

