% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgSyncDBtoRepo.R
\name{uncompressedSize}
\alias{uncompressedSize}
\title{get size of uncompressed .gz file from 4-byte footer}
\usage{
uncompressedSize(gzfile)
}
\arguments{
\item{gzfile}{character scalar filename}
}
\value{
the signed integer taken from the last 4 bytes
of the file (little-endian).  This only represents the
size of the last complete gz stream in the file, and if
the file does nto end with a completed gz stream, then
the value returned by this function is invalid.

i.e. this function's return value is only valid if \code{compressedFileDone(gzfile) == TRUE}
}
\description{
get size of uncompressed .gz file from 4-byte footer
}
