% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMetadataForTags.R
\name{updateMetadataForTags}
\alias{updateMetadataForTags}
\title{update metadata for a set of tags}
\usage{
updateMetadataForTags(t, meta, p = meta("select * from projs"),
  fixBI = FALSE)
}
\arguments{
\item{t:}{data.frame of tags, as returned by motusSearchTags()}

\item{meta:}{safeSQL object to metadata DB}

\item{p:}{data.frame of projects, as returned by motusListProjects()
default: meta("select * from projs")}

\item{fixBI:}{logical scalar; should burst intervals be corrected?
default: FALSE}
}
\value{
TRUE on success.  Generates an error on failure.
}
\description{
These tags might be newly registered, or grabbed from motus.org
when refreshing the metadata cache.  Either way, they might be new
tags and/or have new or modified deployments.  After cleaning up
registration and deployment info for the tags, this function
updates the metadata cache (including the derived \code{events}
table) and its slim copy in the Motus mariaDB, and commits any
changes to the metadata-history repo.
}
\note{
: this function \emph{must} be called within an EXCLUSIVE transaction on
    \code{meta}.  This is to ensure that changes to the \code{tags} and
    \code{events} tables are atomic with respect to the tag finder
    (`find_tags_motus`) program, so that the latter can record the
    commit hash from the metadata-history that corresponds to the
    metadata it is using.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
