% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTagDeployments.R
\name{updateTagDeployments}
\alias{updateTagDeployments}
\title{Pull any new or modified tag deployment records from motus}
\usage{
updateTagDeployments()
}
\value{
TRUE if any tag metadata were updated; FALSE otherwise
}
\description{
This locks the metadata DB before querying motus for any changes to tag
deployment metadata since the 'mtime' record in the 'meta' table
of that DB.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
