% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateEmail.R
\name{validateEmail}
\alias{validateEmail}
\title{Verify that an incoming data email is valid, and return the name of
the user who sent it.}
\usage{
validateEmail(msg)
}
\arguments{
\item{msg}{the full text of the email.}
}
\value{
if the message contains a valid token, returns a 3-element
    list with these items:
\itemize{
\item username the sensorgnome.org username who owns the token
\item email the email address of the user
\item expired logical scalar; TRUE iff token the token has expired
}

Otherwise, returns NULL.
}
\description{
The body and text of the email is searched for a valid token (see
\link{\code{getUploadToken}}) and if found, the username and email
address are returned.
}
\seealso{
\link{\code{getUploadToken}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
