% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateMotusTagDB.R
\name{validateMotusTagDB}
\alias{validateMotusTagDB}
\title{Perform sanity checks on the motus tag database, reporting problems.}
\usage{
validateMotusTagDB()
}
\value{
a list of problems found.
}
\description{
Mainly, look for duplicated tags.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

