% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavFindTags.R
\name{wavFindTags}
\alias{wavFindTags}
\title{detect tags in a .wav audio file}
\usage{
wavFindTags(f, tdb, pd = "findpulsetdbatch", pdpars = list(minsnr = 8,
  minfreq = 0, maxfreq = 24, plen = 2.5), tfpars = list(min_dfreq = -24,
  max_dfreq = 24, signal_slop = 30, frequency_slop = 0.2, pulses_to_confirm = 4,
  pulse_slop = 1.5, max_skipped_bursts = 3, default_freq = 166.376,
  unsigned_dfreq = 1), maxFreqSD = 0.2)
}
\arguments{
\item{f}{path to the .wav file}

\item{tdb}{path to the tag database file}

\item{pd}{character scalar; which pulse detector to use.  One of
c("findpulsefdbatch" or "findpulsetdbatch").  Names one of the
VAMP plugins in the library lotek-plugins.so  Default: "findpulsetdbatch"}

\item{pdpars}{named list of parameters to the pulse detector}

\item{tfpars}{named list of parameters to the tag finder.}

\item{maxFreqSD}{maximum freqsd of returned detections; default: 0.2
kHz}
}
\value{
a data.frame of tag detections with these columns:
\itemize{
\item ts timestamp; seconds since start of file
\item id; id from the database in \code{tdb}
\item dfreq; frequency offset, in kHz
\item sig; signal strength in dB max
}
}
\description{
A .wav file is processed to detect tag pulses, and these are run through
a simplified version of the tag finder to detect individual bursts from
Lotek coded ID tags.
}
\details{
This uses the same signal processing plugin as runs on the sensorgnome,
but via a stripped down vamp-host (the version on the sensorgnome
handles live input from attached funcubes).
}
\note{
uses external programs vamp-host and find_tags_unifile, and
library lotek-plugins.so
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
