% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{authenticate_user}
\alias{authenticate_user}
\title{authenticate_user return a list of projects and receivers the user is authorized to receive data for}
\usage{
authenticate_user(env)
}
\arguments{
\item{user}{motus user name}

\item{password}{motus password (plaintext)}
}
\value{
a JSON list with these items:
\itemize{
\item token character scalar token used in subsequent API calls
\item expiry numeric timestamp at which \code{token} expires
\item userID integer user ID of user at motus
\item projects list of projects user has access to; indexed by integer projectID, values are project names
\item receivers FIXME: will be list of receivers user has access to
}
if the user is authorized.  Otherwise, return a JSON list with a single item
called "error".
}
\description{
This is an app used by the Rook server launched by \code{\link{dataServer}}
Params are passed as a url-encoded field named 'json' in the http POST request.
The return value is a JSON-formatted string
}

