% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{batches_for_tag_project}
\alias{batches_for_tag_project}
\title{get batches for a tag project}
\usage{
batches_for_tag_project(env)
}
\arguments{
\item{projectID}{integer project ID}

\item{ts}{numeric timestamp}
}
\description{
get batches for a tag project
}

