% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{remove old processed data from a receiver database.}
\usage{
cleanup(src, dropTables = FALSE, vacuum = FALSE)
}
\arguments{
\item{src}{dplyr src_sqlite to receiver database}

\item{dropTables}{boolean; if TRUE, erase all output tables, rather
than just emptying them.  This makes sense if their schema
has changed, but is usually not required, so the default is FALSE.}

\item{vacuum}{boolean; if TRUE (the default), free unused storage
from the database.  This can be slow, as the entire database is
rewritten.}
}
\description{
This deletes all output and record of processing.  Original
raw data is not affected.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
