% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropBatchesFromTransfer.R
\name{dropBatchesFromTransfer}
\alias{dropBatchesFromTransfer}
\title{delete a set of batches and all related data from the master data base}
\usage{
dropBatchesFromTransfer(batchIDs)
}
\arguments{
\item{batchIDs}{integer vector of batch IDs}
}
\value{
integer vector of those batchIDs which were actually deleted.
}
\description{
delete a set of batches and all related data from the master data base
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
