% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{email}
\alias{email}
\title{Send an email message.}
\usage{
email(to, subj, msg, ...)
}
\arguments{
\item{to}{email address of recipient}

\item{subj}{character scalar subject line}

\item{msg}{character scalar message content.  This is treated as a
\code{sprintf}-style formatting string, with fields filled in
from \code{...}.}

\item{...}{paramters for replacing \code{sprintf} formatting codes
in \code{msg}}
}
\value{
invisible(NULL)
}
\description{
An email message is sent from data@sensorgnome to the specified user.
A log entry is made, and the outgoing message is saved in the motus
server outbox.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{sprintf}} for formatting codes.
}

