% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailServer.R
\name{emailServer}
\alias{emailServer}
\title{process incoming emails}
\usage{
emailServer(tracing = FALSE)
}
\arguments{
\item{tracing;}{default FALSE.  If TRUE, enter debug browser before
calling each handler.}
}
\value{
This function does not return; it is meant for use in an R
    script run in the background.  After each subjob is handled,
    the function checks for the existence of a file called
    \code{MOTUS_PATH$INBOX/killE}.  If that file is found,
    the function calls quit(save="no").  The file will also
    be detected within the call to feed() when the queue
    is empty, because it is located in the watched folder.
}
\description{
Watch for new messages in \code{MOTUS_PATH$INBOX}.
}
\details{
When a new message is found:
\itemize{
\item create a new job folder in \code{MOTUS_PATH$EMAIL_QUEUE}
\item unpack the email's parts (e.g. attachments, or enclosed forwarded messages)
\item validate by looking for an authorization token
\item save attachments
\item download files from any links
\item run basic sanity checks on files
\item unpack archives
\item email the sender with an acknowledgement and pointer to a status page.
\item enqueue a new job with all files
}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
