% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enqueue.R
\name{enqueue}
\alias{enqueue}
\title{Add a file or folder to the server queue.}
\usage{
enqueue(path, part1, ...)
}
\arguments{
\item{path}{the full path to the file or directory.}

\item{part1}{if present, a string to add to the name of the
new path.}

\item{...}{if present, additional strings to add to the name
of the new path.}
}
\value{
TRUE on success; FALSE otherwise
}
\description{
Add a file or folder to the server queue.
}
\details{
Items specified in \code{part1} and \code{...} are appended
to the path, separated by '_' (underscore).

After this call, the file or directory will no longer exist at the
same location, but will be moved into the server's incoming
directory.
}
\examples{
## Not run:
enqueue("/tmp/mytmpurlfile")

}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{server}}
}

