% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enqueue.R
\name{enqueue}
\alias{enqueue}
\title{Add a file or folder to the server queue.}
\usage{
enqueue(path)
}
\arguments{
\item{path}{the full path to the new file or directory

After this call, the file or directory will no longer exist at the
same location.}
}
\value{
TRUE on success; FALSE otherwise
}
\description{
Add a file or folder to the server queue.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{server}}
}

