% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureMotusTransferTables.R
\name{ensureMotusTransferTables}
\alias{ensureMotusTransferTables}
\title{Ensure we have the mysql transfer tables for moving SG data to
motus.}
\usage{
ensureMotusTransferTables(recreate = c())
}
\arguments{
\item{recreate}{vector of table names which should be dropped then re-created,
losing any existing data.  Defaults to empty vector, meaning no tables
are recreate.}
}
\description{
Ensure we have the mysql transfer tables for moving SG data to
motus.
}
\note{
: assumes \link{\code{openMotusDB()}} has been called, so that
the global variable MotusDB is set.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
