% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureServerDB.R
\name{ensureServerDB}
\alias{ensureServerDB}
\title{ensure we have created the server database}
\usage{
ensureServerDB(installing = FALSE)
}
\arguments{
\item{installing;}{logical scalar; if TRUE, the caller is part of
package installation, rather than a running server, and the locking
of the `ServerDB` symbol is skipped.  Default:  FALSE}
}
\value{
no return value, but saves a safeSQL connection to the server database
in the global symbol \code{ServerDB}
}
\description{
This database holds a table of symbolic locks (e.g. to prevent multiple
processes from running a job on the same receiver in parallel)
It also holds all job information in table \code{jobs}, but that
table is ensured by the call to \link{\code{Copse()}} in \link{\code{loadJobs()}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
