% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixDOSfilenames.R
\name{fixDOSfilenames}
\alias{fixDOSfilenames}
\title{Correct DOS-style 8.3 filenames into (almost) their original full filenames.}
\usage{
fixDOSfilenames(f, info, onDisk = FALSE)
}
\arguments{
\item{f}{vector of full paths to files}

\item{info}{data frame of split filename components, which are the
named capture groups in \link{\code{sgFilenameRegex}}}

\item{onDisk;}{logical scalar; if TRUE, also rename the file on disk
default: FALSE}
}
\value{
info, with corrections to those rows representing DOS-style
    filenames whose receiver can be deduced.
}
\description{
Any files needing renaming are renamed on disk. The corrected
filename will have the receiver serial number, the first timestamp
from the file, and a boot number of 0.  This flags the requirement
to correct the boot number after other files have been merged into
their receiver databases.
}
\details{
Some users move SG data files along a route that fails to
    preserve their long filenames, forcing them into a DOS-style
    8.3 filename.  Typically, this only happens to a few files per
    batch, for some reason.  Perhaps they are storing raw files in
    the top level folder of a VFAT filesystem, which limits the
    total length of filenames in a folder.

In any case, these can be recognized by a tilde \code{~} in the
filename, and the first two letters of the shortened name will
match (case insensitively) the first two letters of the original
full name.
}
\note{
this function should be made trivial in the future by writing
    the bootnum and serial number at the start of each file.

Corrections require inferring the serial number and boot
    count for any 8.3 filename, and then using the timestamp from
    the first record in the file as the file timestamp.

\itemize{
\item for each 8.3 filename, we use as context the file's folder

\item if the first two characters of the
shortened name match (case insensitively) the
first two characters of any file with unshortened name in the context.
If so, and if the matching full-named files all have the same serial
number, correct appropriately.

\item otherwise, we can't tell what receiver the file belongs to,
and we leave its receiver column as NA. FIXME: get first and last
timestamps in shortened files and compare to timestamps parsed from
full names.

}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

