% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDownloadURL.R
\name{getDownloadURL}
\alias{getDownloadURL}
\title{return the URL to the download site for a project}
\usage{
getDownloadURL(projectID)
}
\arguments{
\item{port}{projectID integer scalar; motus project ID}
}
\value{
character scalar giving URL of download site
}
\description{
return the URL to the download site for a project
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
