% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjDir.R
\name{getProjDir}
\alias{getProjDir}
\title{Get the directory where files for a project are stored.}
\usage{
getProjDir(p)
}
\arguments{
\item{p}{integer scalar; motus project ID}
}
\value{
character scalar; the path to the project directory.
}
\description{
Get the directory where files for a project are stored.
}
\note{
As a side-effect, the project directory will be created and populated
with appropriate files (e.g. header.html) if it does not already exist.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

