% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYearProjSite.R
\name{getYearProjSite}
\alias{getYearProjSite}
\title{Get the project and site folders, given a year and receiver serial number.}
\usage{
getYearProjSite(serno, ts = NULL, bootnum = NULL)
}
\arguments{
\item{serno}{character vector of serial numbers}

\item{ts}{numeric vector with non-NA entries wherever \code{serno}
is a Lotek receiver. Can be \code{NULL}, the default, if all
\code{serno} are SensorGnomes.}

\item{bootnum}{integer vector with non-NA entries wherever
\code{serno} is a SensorGnome.  Can be \code{NULL}, the
default, if all \code{serno} are Lotek receivers.}
}
\value{
a data.frame with these columns:
\itemize{
\item serno receiver serial number
\item year integer
\item proj project folder in /SG/year, or NA if none found
\item site site subfolder in /SG/year/proj, or NA if none found
\item tsStart timestamp at which this year/proj/site begins (or NA if SG)
\item bootnumStart boot session at which this year/proj/site begins (or NA if Lotek)
}
}
\description{
This is a deprecated function to help process data files in the old
sensorgnome YEAR/PROJECT/SITE hierarchy.  Sites are looked up
by bootnum for sensorgnomes, and by timestamp for Lotek receivers.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

