% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYearProjSite.R
\name{getYearProjSite}
\alias{getYearProjSite}
\title{Get project and site names, given receiver serial numbers and timeranges}
\usage{
getYearProjSite(serno, tsLo, tsHi)
}
\arguments{
\item{serno}{character vector of serial numbers}

\item{tsLo}{numeric vector of start times}

\item{tsHi}{numeric vector of end times}
}
\value{
a data.frame with these columns:
\itemize{
\item serno receiver serial number
\item year integer year deployment began
\item proj name of project
\item site name of receiver deployment
\item tsStart timestamp at which this deployment begins
\item tsEnd timestamp at which this deployment ends
}
}
\description{
This is meant to provide project and site labels for plots and summary files.
For each receiver in the input, all deployments that overlap the specified
time range are returned.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

