% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{gps_for_tag_project}
\alias{gps_for_tag_project}
\title{get all GPS fixes from a batch "near" to detections of tags from a project.}
\usage{
gps_for_tag_project(env)
}
\arguments{
\item{projectID}{integer project ID of tags of interest}

\item{batchID}{integer batchID}

\item{ts}{numeric timestamp of latest fix already obtained}
}
\value{
a data frame with the same schema as the gps table, but JSON-encoded as a list of columns
}
\description{
get all GPS fixes from a batch "near" to detections of tags from a project.
}
