% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hackError.R
\name{hackError}
\alias{hackError}
\title{examine the calling environments of a failed job}
\usage{
hackError(j, type, index)
}
\arguments{
\item{j}{the top-level motus job ID one of whose subjobs failed}

\item{type}{the type of subjob to examine.  If not specified,
the first such subjob is examined.}
}
\description{
examine the calling environments of a failed job
}
