% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleDownloadableLink.R
\name{handleDownloadableLink}
\alias{handleDownloadableLink}
\title{Grab content from a downloadable link.}
\usage{
handleDownloadableLink(path, isdir)
}
\arguments{
\item{path}{the full path to the file with the download link}

\item{isdir}{boolean; TRUE iff the path is a directory}
}
\value{
TRUE if \code{path} contained a link which was
    successfully downloaded.

If the content is successfully downloaded, it is enqueued.
}
\description{
The link is contained a file whose name looks like url_XXX.txt
The file contains a single line, with this form:
}
\details{
\code{TYPE URL}

where TYPE is the name of a handler, e.g. dropbox for which a function
\code{dowload.TYPE} is defined, and \code{URL} is the location
of the item.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

