% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleDTA.R
\name{handleDTA}
\alias{handleDTA}
\title{handle a folder of .DTA files}
\usage{
handleDTA(j)
}
\arguments{
\item{j, }{the job.}
}
\value{
TRUE after queueing jobs
}
\description{
Called by \code{\link{processServer}} for a file or folder added
to the queue.  Merges files into receiver DBs, then queues
a job to run the tag finder on each of these.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{processServer}}
}

