% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleEmail.R
\name{handleEmail}
\alias{handleEmail}
\title{handler for incoming email}
\usage{
handleEmail(path, isdir, params)
}
\arguments{
\item{path}{the full path to the new file or directory}

\item{isdir}{boolean; TRUE iff the path is a directory}

\item{params}{not used}
}
\value{
TRUE if the file is an email; FALSE otherwise
}
\description{
called by \code{\link{server}} for emails.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{server}}
}

