% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleEmail.R
\name{handleEmail}
\alias{handleEmail}
\title{handler for new incoming email}
\usage{
handleEmail(j)
}
\arguments{
\item{j}{the job; it has this field:
\itemize{
\item msgFile: full path to message file
}}
}
\value{
TRUE if the email message was successfully handled
}
\description{
called by \code{\link{emailServer}} for new messages
}
\seealso{
\code{\link{emailServer}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
