% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleExportData.R
\name{handleExportData}
\alias{handleExportData}
\title{push newly generated data to motus transfer tables}
\usage{
handleExportData(j)
}
\arguments{
\item{j}{the job, with these fields:
\itemize{
\item serno - the receiver serial number
}}
}
\value{
TRUE
}
\description{
Any new batches for this receiver, and their associated runs, hits,
and GPS fixes are added to motus transfer tables.
}
\note{
if the top level job has a TRUE value for parameter
    \code{isTesting}, then the batches transferred will be given
    \code{status = -1}, rather than the usual {status = 1}.  This tells
    \code{\link{dataServer()}} not to return records for such
    batches unless explicitly requested to by the \code{includeTesting} parameter.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
