% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleLogs.R
\name{handleLogs}
\alias{handleLogs}
\title{Save a folder of linux system log files from a sensorgnome.}
\usage{
handleLogs(j)
}
\arguments{
\item{j}{the job}
}
\value{
TRUE iff all files in the folder could be archived; i.e. iff a valid
    sensorgnome serial number was found in at least one of the log
    files
}
\description{
Called by \code{\link{processServer}} for a folder containing
a file called syslog
}
\note{

}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{server}}
}

