% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleLtFindtags.R
\name{handleLtFindtags}
\alias{handleLtFindtags}
\title{run the tag finder on all files from a Lotek receiver}
\usage{
handleLtFindtags(j)
}
\arguments{
\item{j, }{the job.  It has these properties:
\enumerate{
\item serno: serial number of receiver; "Lotek-NNN"
\item tsStart: earliest timestamp of a detection in a file
with new data.  Not used at present, but might
allow for pause/resume later.
}}
}
\value{

}
\description{
Called by \code{\link{processServer}}. After running the tag
finder, queues a new subjob that exports data.
}
\note{
FIXME: we run all data from the receiver in a single go; we need
a batch concept for these receivers, if for no other reason than
to allow time/deployment-dependent parameter overrides.
see \link{https://github.com/jbrzusto/motusServer/issues/60}
For now, any project or receiver parameter override active
at the time the function is called will be used.

if topJob(j) contains fields motusUserID and motusProjectID, these are
recorded in the receiver database for the batch generated by this call.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{processServer}}
}

