% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleOldExport.R
\name{handleOldExport}
\alias{handleOldExport}
\title{export data the 'old' (pre-motus) way.}
\usage{
handleOldExport(j)
}
\arguments{
\item{j}{the job, with these fields:
\itemize{
\item serno - the receiver serial number
\item monoBN - the range of receiver bootnums; NULL for Lotek receivers.
\item ts - the approximate range of timestamps; NULL for SGs.
}}
}
\value{
TRUE if the export succeeded
}
\description{
export data the 'old' (pre-motus) way.
}
\details{
We generate Year/Proj/Site plots for the receiver, showing
    hourly tag detections and antenna status, then upload these to
    the user's wiki page at sensorgnome.org; a copy or symlink
    is placed in /sgm/plots
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

