% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleRegisterTags.R
\name{handleRegisterTags}
\alias{handleRegisterTags}
\title{handle a batch of tag registrations}
\usage{
handleRegisterTags(j)
}
\arguments{
\item{j}{the job}
}
\value{
TRUE;
}
\description{
Tag registrations are recordings of tag output accompanied by a
text file of metadata.  This function processes the recordings and
pushes the valid ones to motus.org
}
\details{
The job folder must contain a file called tagreg.txt with these lines:

\itemize{
\item  motusProjID:  XXX (numeric project ID)
\item  tagModel: NTQB-??? (as given on the Lotek invoice)
\item  nomFreq: 166.38 (nominal tag frequency, in MHz)
\item  species: XXXXX (optional 4-letter code or motus numeric species ID)
\item  deployDate: YYYY-MM-DD (earliest likely deployment date for any tags)
\item  codeSet: X (optional codeset ; default: 4 for "Lotek4"; can also be 3 for "Lotek3")
}

as well as one or more recording files with names like \code{tagXXX.wav}
where \code{XXX} is the manufacturer's tag ID, typically 1 to 3 digits.
When there are recordings of tags with the same ID but different burst intervals,
the 2nd, 3rd, and so on such tags are given names like \code{tagXXX.1.wav, tagXXX.2.wav, ...}
}
\note{
By default, we assume each tag was recorded at 4 kHz below
    its nominal frequency; e.g.  at 166.376 MHz for a nominal
    166.38 MHz tag.  If that's not true, the filename should
    include a portion of the form \code{@XXX.XXX} giving the
    frequency at which it was recorded;
    e.g. \code{tag134@166.372.wav} indicates a tag recorded at
    166.372 MHz, rather than the default.

2017-06-06 ANTC-M tags will be registered as NTQB- models like so:
    ANTC-M1-1     NTQB-1
    ANTC-M2-1     NTQB-2
    ANTC-M3-2     NTQB-3-2
    ANTC-M4-2S    NTQB-4-2
    ANTC-M4-2L    NTQBW-4-2
    ANTC-M6-1     NTQB-6-1
    ANTC-M6-2     NTQB-6-2
but with an entry appended to /sgm/tagnotes.txt giving the database query required on motus.org
to correct the model once it is supported there.
Called by \code{\link{processServer}}.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\code{\link{processServer}}, which calls this function.
}

