% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleRerunReceiver.R
\name{handleRerunReceiver}
\alias{handleRerunReceiver}
\title{rerun data from a receiver}
\usage{
handleRerunReceiver(j)
}
\arguments{
\item{j}{the job with these items:

\itemize{

\item serno character scalar; the receiver serial number

\item monoBN integer vector of length 2; the range of receiver boot
sessions to run; NULL for Lotek receivers.  If not specified for an
SG receiver, then all boot sessions are rerun.

}}
}
\value{
TRUE
}
\description{
Called by \code{\link{processServer}}
}
\details{
reruns the tagfinder for a receiver, possibly limiting the
re-run to one or more boot sessions (for SGs).
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{processServer}}
}

