% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleRerunReceiver.R
\name{handleRerunReceiver}
\alias{handleRerunReceiver}
\title{re-run data from a receiver}
\usage{
handleRerunReceiver(j)
}
\arguments{
\item{j}{the job with these items:

\itemize{

\item serno character scalar; the receiver serial number

\item monoBN integer vector of length 2; the range of receiver boot
sessions to run; If not specified, then for an SG receiver, all
boot sessions are rerun.  Ignored for Lotek receivers.
If only exporting, then export from the latest deployment record
that is not later than \code{monoBN[1]}.

\item ts real vector of length 1 or 2; the start (and possibly end)
timestamp of boot sessions to run.  Ignored for an SG. If not
specified for a Lotek receiver, exports all data.  Otherwise, exports
data only in the specified period (if length is 2) or starting
at the specified timestamp (if length is 1)

\item exportOnly logical scalar; if TRUE, skip running the tagfinder

\item cleanup logical scalar; if TRUE, delete all hits, runs,
batches before re-running.  Ignored (and treated as FALSE) if
monoBN is specified.  i.e. cleanup can only be specified for a
full re-run.

\item [optional] paramOverrides; if present, must be a string scalar
specifying parameters to the tag finder, which will be appended to
any default and registered override values; thus, parameter values specified
here will override those from other sources.  This item is handled
in \code{\link{handleSGfindtags}} or \code{\link{handleLtFindtags}}
}}
}
\value{
TRUE
}
\description{
Called by \code{\link{processServer}}
}
\details{
re-runs the tagfinder for a receiver, possibly limiting the
re-run to one or more boot sessions (for SGs).
}
\seealso{
\link{\code{processServer}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
