% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleServerFiles.R
\name{handleServerFiles}
\alias{handleServerFiles}
\title{handle a folder of files from the server}
\usage{
handleServerFiles(j)
}
\arguments{
\item{j}{the job
The job can have these properties:
\itemize{
\item sanityCheck if TRUE, sanity check files, which is
time-consuming but a good idea if files come from a source where they might
not have been fully sync'd (e.g. dropbox) before the user sent us a link to them.
}}
}
\value{
TRUE; As a side effect, subjobs for sanity checking files and
unpacking archives are queued.
}
\description{
Called by \code{\link{processServer}}.  Any files in subfolders
of the specified job's folder are processed.  These might be archives
that need to be sanity checked and unpacked.
}
\seealso{
\code{\link{processServer}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
