% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleSGfindtags.R
\name{handleSGfindtags}
\alias{handleSGfindtags}
\title{handle the job of running the tag finder on one sensorgnome boot session}
\usage{
handleSGfindtags(j)
}
\arguments{
\item{j}{the job, with these properties:
\itemize{
\item serno serial number of receiver (including leading 'SG-')
\item monoBN monotonic boot session number
\item boolean TRUE or FALSE; can the tag finder be run with \code{--resume}?
}}
}
\value{
TRUE on success, or FALSE if the tag finder has an error.
}
\description{
Called by \code{\link{processServer}}.
}
\details{
Runs the tag finder for one boot session of a single SG.
The files for that boot session have already been merged into
the DB for that receiver.
}
\note{
if topJob(j) contains fields motusUserID and motusProjectID, these are
recorded in the receiver database for the batch generated by this call.
}
\seealso{
\code{\link{server}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
