% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleSyncReceiver.R
\name{handleSyncReceiver}
\alias{handleSyncReceiver}
\title{sync from an attached receiver, processing any new data}
\usage{
handleSyncReceiver(j)
}
\arguments{
\item{j}{the job with these items:

\itemize{

\item serno character scalar; the receiver serial number

}}
}
\value{
TRUE
}
\description{
Called by \code{\link{syncServer}}
}
\details{
grabs any new data from the attached receiver, and
re-runs the tagfinder
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{syncServer}}
}

