% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobLog.R
\name{jobLog}
\alias{jobLog}
\title{Add a message to the log for a job.  Log messages are saved in two
JSON fields called "log" and "summary" in the top-level ancestor of the job.
"Summary" is intended to give a quick overview, while "log" should provide.
details.}
\usage{
jobLog(j, msg, summary = FALSE)
}
\arguments{
\item{j}{the job}

\item{msg}{character vector of messages}

\item{summary}{logical scalar; if TRUE, the message is added to the summary field; otherwise,
the default, it is added to the log field.


The messages in \code{msg} are joined with "\\n" and appended to any already existing for this job.}
}
\description{
Add a message to the log for a job.  Log messages are saved in two
JSON fields called "log" and "summary" in the top-level ancestor of the job.
"Summary" is intended to give a quick overview, while "log" should provide.
details.
}

