% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobProduced.R
\name{jobProduced}
\alias{jobProduced}
\title{report and maybe record what a job produced}
\usage{
jobProduced(j, u)
}
\arguments{
\item{j}{job integer number.  If not a topjob, its topjob is used instead.}

\item{u}{character vector of job product URLS.  Can be missing.}
}
\value{
The character vector of job product URLs, after adding any specified
by \code{u}.
}
\description{
Most jobs generate or modify files that the user can access via URL.
This function returns a vector of product URLs, after adding any new
ones specified.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
