% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadJobs.R
\name{loadJobs}
\alias{loadJobs}
\title{load existing jobs from server database}
\usage{
loadJobs(stumpType)
}
\arguments{
\item{stumpType}{only examine jobs whose stump is of the specified type.}
}
\value{
no return value.
}
\description{
A global object \code{Jobs} is created, which manages jobs.
It is populated from the "jobs" table in the server database.
}
\details{
It has these fixed fields:
\itemize{

\item type: character; type of job

\item done: integer; 0 if not attempted; negative for errors; +1
for successfully finished

\item path: full path to folder for job

\item oldpath: full path to previous folder for job (permits
recovery from server failure during job move)

}

Other job items are stored in a JSON-encodded text field called
\code{data}.

For any jobs which are not done and which are of the specified
type, we verify that the "path" field is correct, in case the
server was interrupted while moving a job.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

