% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadJobs.R
\name{loadJobs}
\alias{loadJobs}
\title{Load unfinished jobs from server database and enqueue them}
\usage{
loadJobs(headType)
}
\arguments{
\item{headType}{only examine jobs whose head job is of this type.}
}
\value{
no return value.  Jobs are enqueued in the global \code{MOTUS_QUEUE}.
}
\description{
A global object \code{Jobs} is created, which manages jobs.
It is populated from the "jobs" table in the server database.
}
\details{
It has these fixed fields:
\itemize{

\item type: character; type of job

\item done: integer; 0 if not attempted; negative for errors; +1
for successfully finished

\item path: full path to folder for job

\item oldpath: full path to previous folder for job (permits
recovery from server failure during job move)

}

Other job items are stored in a JSON-encodded text field called
\code{data}.

For any jobs which are not done and whose head job is of the specified
type, we verify that the "path" field is correct, in case the
server was interrupted while moving a job.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

