% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggingTry.R
\name{loggingTry}
\alias{loggingTry}
\title{Try evaluate expression, logging the call stack in case of an error.}
\usage{
loggingTry(j, expr)
}
\arguments{
\item{expr}{expression to evaluate}

\item{job, }{to which any error message will be looged}
}
\value{
value of \code{expr}
}
\description{
This function wraps handlers called by \link{\code{server}}.  When
an error occurs, the timestamped stack dump is saved in
\code{MOTUS_PATH$errors}
}
\note{
based on R's \code{util::dump.frames()} and \code{tools:::.try_quietly()}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
