% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltFindTags.R
\name{ltFindTags}
\alias{ltFindTags}
\title{find tags in a Lotek receiver dataset}
\usage{
ltFindTags(src, tagDB, par = NULL)
}
\arguments{
\item{src}{dplyr src_sqlite to (lotek) receiver database}

\item{tagDB}{path to sqlite tag registration database}

\item{par}{list of parameters to the filtertags code.  Defaults
to NULL;}
}
\value{
a list with these items
\itemize{
  \item batchID the batch number
  \item numHits the number of unfiltered tag detections in the stream.
}
}
\description{
This function searches for sequences of tag detections
corresponding to registered (ID, burst interval) pairs and adds
them to the hits, runs, batches etc. tables in the receiver
database.
}
\details{
the tag finder is run once for each boot session that
has at least one detection record.  Boot session numbers
are first generated from all the boottime records.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
