% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltMergeFiles.R
\name{ltMergeFiles}
\alias{ltMergeFiles}
\title{Merge a batch of Lotek .DTA files with an existing database.}
\usage{
ltMergeFiles(files, dbdir = MOTUS_PATH$RECV)
}
\arguments{
\item{files}{either a character vector of full paths to files, or the full
path to a directory, which will be searched recursively for .DTA files.}

\item{dbdir}{path to folder with existing receiver databases; defaults
to \code{MOTUS_PATH$RECV}}
}
\value{
a data_frame reporting the fate of each file, with these columns:
\enumerate{
\item fullname - full path to filename
\item nameNew  - TRUE iff file of this name was not yet in database
\item dataNew  - TRUE iff file with these contents was not yet in database
\item use      - TRUE iff file copied to database
\item err      - if not NA, any error message generated by trying to parse the file
}
}
\description{
Merge a batch of Lotek .DTA files with an existing database.
}
\note{
If err is not NA for a file, then other fields for that file
    might not be set appropriately in the return value.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

