% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{makeBody}
\alias{makeBody}
\title{make the body for a reply
converts \code{...} to json using toJSON with options:
\itemize{
\item auto_unbox=TRUE
\item dataframe="columns"
}
then compresses the result
with memCompress and method "bzip2"}
\usage{
makeBody(...)
}
\arguments{
\item{...}{items and options for \code{toJSON}}
}
\value{
a raw vector
}
\description{
make the body for a reply
converts \code{...} to json using toJSON with options:
\itemize{
\item auto_unbox=TRUE
\item dataframe="columns"
}
then compresses the result
with memCompress and method "bzip2"
}
