% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusLog.R
\name{motusLog}
\alias{motusLog}
\title{Record messages to the motus main log.}
\usage{
motusLog(fmt, ...)
}
\arguments{
\item{fmt}{character scalar \code{sprintf}-style formatting string}

\item{...}{any parameters required for \code{fmt}}
}
\value{
invisible(NULL)
}
\description{
Works like sprintf, but sends output to the motus main log file,
after prepending the current date/time.  Output is followed by an
end-of-line.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

