% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusRegisterReceiver.R
\name{motusRegisterReceiver}
\alias{motusRegisterReceiver}
\title{register a receiver with motus (admin only)}
\usage{
motusRegisterReceiver(serno, secretKey = NULL)
}
\arguments{
\item{serno:}{character vector, such as "SG-1234BBBK5678"}

\item{secretKey:}{if NULL, generate a pair of public/private ssh keys
for the receiver, and use the SHA1-sum of the private key file as the
secretKey.  Otherwise, this is an upper case hex string giving the
secretKey for the receiver.}
}
\value{
the query result, which won't usually be useful
}
\description{
register a receiver with motus (admin only)
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
