% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusSearchTags.R
\name{motusSearchTags}
\alias{motusSearchTags}
\title{return a list of all motus tags}
\usage{
motusSearchTags(projectID = NULL, tsStart = NULL, tsEnd = NULL,
  searchMode = c("startsBetween", "overlaps"), mfgID = NULL,
  status = NULL, ...)
}
\arguments{
\item{projectID:}{integer scalar; motus internal project ID}

\item{tsStart:}{numeric scalar; start of active period}

\item{tsEnd:}{numeric scalar; end of active period}

\item{searchMode:}{character scalar; type of search
desired. "overlaps" looks for tags active during at least a
portion of the time span \code{c(tsStart, tsEnd)}, while
"startsBetween" looks for tags with deployment start times in
the same range.}

\item{mfgID:}{character scalar; typically a small integer; return
only records for tags with this manufacturer ID (usually
printed on the tag)}

\item{status:}{integer; if non-NULL, returns only tags with the
specified status.  1L = tag finished; 2L = tag active; 0L = tag
not yet deployed}

\item{...:}{additional parameters to motusQuery()}
}
\value{
the data.frame of motus tags and their meta data satisfying
    the search criteria, or NULL if there are none.  The data.frame
    has these columns:

\itemize{
   \item tagID
   \item projectID
   \item mfgID
   \item dateBin
   \item type
   \item codeSet
   \item manufacturer
   \item model
   \item lifeSpan
   \item nomFreq
   \item offsetFreq
   \item period
   \item periodSD
   \item pulseLen
   \item param1
   \item param2
   \item param3
   \item param4
   \item param5
   \item param6
   \item param7
   \item param8
   \item tsSG
   \item approved
   \item deployID
   \item status
   \item tsStart
   \item tsEnd
   \item deferSec
   \item speciesID
   \item markerNumber
   \item markerType
   \item latitude
   \item longitude
   \item elevation
   \item comments
}
}
\description{
return a list of all motus tags
}
\note{
As of 2018-08-24, filtering by `tsStart` and `tsEnd` fails to return any tags
whose `tsEnd` is NULL; see https://github.com/MotusDev/MotusAPI/issues/8
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
