% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveFilesUniquely.R
\name{moveFilesUniquely}
\alias{moveFilesUniquely}
\title{Move files to a single folder, renaming new files in cases
of conflict.}
\usage{
moveFilesUniquely(src, dst)
}
\arguments{
\item{src}{character vector of full paths of files to move}

\item{dst}{path to target folder}
}
\value{
a character vector of the same length as \code{src}, with non-NA
entries giving new names for any files that were renamed
}
\description{
Move files to a single folder, renaming new files in cases
of conflict.
}
\details{
Filenames that conflict with each other or those already in
\code{dst} are changed like so:

\enumerate{
\item if the filename does not end in \code{-NNN}  where \code{NNN} is an
integer, then add \code{-1} to
the filename; e.g.

   \code{myfile.txt -> myfile.txt-1}

\item if the filename already ends in \code{-NNN}, then
increment \code{NNN}; e.g.

  \code{myfile.txt-3 -> myfile.txt-4}

}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
