% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveFilesUniquely.R
\name{moveFilesUniquely}
\alias{moveFilesUniquely}
\title{Move files to a single folder, renaming new files in cases
of conflict.}
\usage{
moveFilesUniquely(src, dst, ...)
}
\arguments{
\item{src}{character vector of full paths of files to move}

\item{dst}{path to target folder}
}
\value{
a boolean vector of the same length as \code{src}, with TRUE
entries corresponding to files moved successfully.
}
\description{
Move files to a single folder, renaming new files in cases
of conflict.
}
\details{
Filenames that conflict with each other or those already in
\code{dst} are changed like so:

\enumerate{
\item if the filename does not end in \code{-NNN.EXT}  where \code{NNN} is an
integer and \code{EXT} is the file extension, then add \code{-1} to
the filename before the extension; e.g.

   \code{myfile.txt -> myfile-1.txt}

\item if the filename already ends in \code{-NNN.EXT}, then
increment \code{NNN}; e.g.

  \code{myfile-3.txt -> myfile-4.txt}

}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

