% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveJob.R
\name{moveJob}
\alias{moveJob}
\title{move a top job to a new folder}
\usage{
moveJob(j, dest)
}
\arguments{
\item{j}{Twig object representing the job.}

\item{dest}{path to new folder where the job's folder will become
a subfolder.}
}
\value{
TRUE on success; FALSE if the job's folder could not be moved

Note: this function can only move a top job.  To move subjobs to different
parents, use \link{\code{reparentJob()}}
}
\description{
Retain the old path to the job, in case of interruption
between recording the new path and moving the folder.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{Copse}}
}

