% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newJob.R
\name{newJob}
\alias{newJob}
\title{create and enqueue a new job}
\usage{
newJob(.type, .parentPath, ..., .parent = NULL, .enqueue = TRUE)
}
\arguments{
\item{.type}{character scalar name of job type}

\item{.parentPath}{the folder in which to create the new job's folder.
If omitted, no folder is created for this job, and the job's \code{$path}
item is set to the \code{.parent$path}, if \code{.parent} is not null.}

\item{...}{additional named metadata for this job}

\item{.parent}{.parent job, if any}

\item{.enqueue}{logical scalar; should the job be added to the current queue?
Default: TRUE, but non-server callers such as scripts should specify
FALSE.}
}
\value{
A Twig (see \link{\code{Copse}}) object \code{j} representing the job.
    The path to the job folder is accessible as \code{j$dir}.
}
\description{
create and enqueue a new job
}
\details{
A job is a Twig object from the Copse created by
\link{\code{loadJobs}}, and possibly an associated folder named NNNNNNNN,
where N is the job number padded on the left with zeroes to a
length of 8 digits.  Files associated with the job are stored in
the folder, while metadata are stored in the Twig object, which is
maintained in persistent storage in the Copse's SQLite database.
Jobs can have subjobs via the Twig's parent() relationship.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{Copse}}
}

