% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newJob.R
\name{newJob}
\alias{newJob}
\title{create a new job.}
\usage{
newJob(type, params)
}
\arguments{
\item{type}{character scalar name of job handler}

\item{params}{R object representable in JSON}
}
\value{
This function returns a named integer vector of length 1
    with class "motusJob".  The value is the job number, and the
    name is the full path to the new folder.
}
\description{
A new unique job # is generated, and a folder for it is added to /sgm/queue/0
An entry in the server database is created.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

