% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openMotusDB.R
\name{openMotusDB}
\alias{openMotusDB}
\title{Return a src_mysql attached to the motus mysql database.}
\usage{
openMotusDB(dbname = "motus", host = "localhost", user = "motus")
}
\arguments{
\item{dbname}{database name; default: "motus"}

\item{host}{hostname or IP address; default: "localhost"}

\item{user}{user on database; default: "motus"}
}
\value{
object of class \link{\code{safeSQL}}
}
\description{
The motus database holds batches of tag detections intended for the
motus server.  This must be run by a user who has permissions to
the mysql motus database on the local server, and whose password to
that database is stored in the file ~/.secrets/motusSecrets as
element "dbPasswd".
}
\note{
The safeSQL objects is stored in the global variable MotusDB, and
    if that variable already exists, the connection it holds is
    used.  This means the DB connection is normally opened at most
    once per session.  It is reopened automatically if it has closed.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

