% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFilenames.R
\name{parseFilenames}
\alias{parseFilenames}
\title{parse sensorgnome filenames into components}
\usage{
parseFilenames(f, base = basename(f), checkDOS = TRUE)
}
\arguments{
\item{f}{character vector of filenames with full path}

\item{base}{character vector of file basenames; these default to
\code{basename(f)}, but will differ when the basename has been
corrected (e.g. to remove invalid UTF-8 sequences).}

\item{checkDOS}{if TRUE, the default, try to correct DOS 8.3-style filenames;
if FALSE, return NA for rows corresponding to these.}
}
\value{
a dataframe of components, with one row per filename and these columns:

\enumerate{
 \item prefix:  human readable short site name
 \item serno:  receiver serial number; "SG-" followed by 12 alphanumeric characters e.g. 1315BBBK2156, or possibly with an appended "_N" where N is 1, 2, ...
 for disentangling serial number collisions.  Alphabetic characters are converted to upper case.
 \item bootnum:  boot count (integer)
 \item tsString: timestamp in YYYY-MM-DDTHH-MM-SS.SSSS format
 \item ts:  timestamp embedded in name (double, with class \code{c("POSIXt", "POSIXct")} )
 \item tsCode:  timestamp code ('P' means before GPS fix, 'Z' means accurate to 1e-6 s, 'Y' to 1e-5s, 'X' to 1e-4s, ..., 'T' to 1s)
 \item port:  character; port label, if this file is associated with a single port (e.g. a .WAV file); "all" if all ports
 \item extension:  character extension of uncompressed file; e.g. ".txt"; lower case
 \item comp:  character; file compression type, if any:  "", or ".gz"; lower case
}
}
\description{
parse sensorgnome filenames into components
}
\note{
:
\itemize{
\item returns NULL if no filenames match regex; otherwise, return value has rows
filled with NA for any filenames not matching the expected form

\item serial number collisions are resolved based on prefix and possibly other
filename components - see the
variable \code{sernoCollisionFixes}.  Resolution occurs by adding a suffix to the \code{serno}
field returned by this function, but does not rename files.
}
}
