% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queueDownloads.R
\name{queueDownloads}
\alias{queueDownloads}
\title{Parse an email message for links to downloadable content, and queue
them for download.}
\usage{
queueDownloads(j, msg)
}
\arguments{
\item{j}{job for which downloads will be subjobs}

\item{msg}{character scalar; the message text.}
}
\value{
the number of unique links for which download subjobs were queued
}
\description{
The links must be of the forms specified in \link{\code{dataTransferRegex}}
We queue a download subjob for each link.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

