% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeSys.R
\name{safeSys}
\alias{safeSys}
\title{call a shell command safely.}
\usage{
safeSys(cmd, ...)
}
\arguments{
\item{cmd}{full path to the executable file (can be a shell script,
for example)}

\item{...}{list of parameters to the command; these are combined
using \code{c()} and then quoted using \link{\code{shQuote()}}}
}
\value{
no return value.
}
\description{
Quotes all parameters for a bash-type shell then runs the command with
them using \link{\code{system2}}
}
\note{

}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

