% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{size_of_update_for_tag_project}
\alias{size_of_update_for_tag_project}
\title{get count of update items for a tag project}
\usage{
size_of_update_for_tag_project(env)
}
\arguments{
\item{projectID}{integer project ID}

\item{batchID}{integer batchID; only batches with larger batchID are considered}
}
\value{
a list with these items:
\itemize{
\item numBatches
\item numRuns
\item numHits
\item numGPS
}
}
\description{
get count of update items for a tag project
}

