% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFile.R
\name{testFile}
\alias{testFile}
\title{Test files for validity.}
\usage{
testFile(files)
}
\arguments{
\item{files}{vector of paths of files to check}
}
\value{
integer vector of same length as \code{files}, with values
from this list:

\enumerate{
\item if the file passes all tests
\item if the file doesn't exist
\item if the file is non-empty but all zeroes
\item if the file has a .gz, .bz2, .7z, .rar, or .zip extension, but fails
the integrity test of the appropriate archiving program
}
}
\description{
Sometimes users send shared links to folders before all files have sync'd,
so these files might not be valid when downloaded.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

