% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{validate_request}
\alias{validate_request}
\title{validate a request by looking up its token, failing gracefully}
\usage{
validate_request(json, res, needProjectID = TRUE, needAdmin = FALSE)
}
\arguments{
\item{json}{named list with at least these items:
\itemize{
\item authToken authorization token
\item projectID (optional) projectID
}}

\item{res}{the rook result object}

\item{needProjectID}{logical; if TRUE, a projectID to which the user
has permission must be in \code{json}; default:  TRUE}

\item{needAdmin}{logical; if TRUE, the user must have userType="administrator"
in order to use the entry point; default:  FALSE}
}
\value{
if \code{authToken} represents a valid unexpired token, and needProjectID is FALSE or projectID is
a project for which the user is authorized, returns
a list with these items:
\itemize{
\item userID integer user ID
\item projects integer vector of *all* project IDs user has permission to
\item projectID the projectID specified in the request (and it is guaranteed the user has permission to this project)
}
Otherwise, send a JSON-formatted reply with the single item "error": "authorization failed"
and return NULL.
}
\description{
validate a request by looking up its token, failing gracefully
}
