% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServer.R
\name{validate_request}
\alias{validate_request}
\title{validate a request by looking up its token, failing gracefully}
\usage{
validate_request(req, res)
}
\arguments{
\item{authToken}{authorization token}

\item{projectID}{(optional)}
}
\value{
if \code{authToken} represents a valid unexpired token, and projectID is
a project for which the user is authorized, returns
a list with these items:
\itemize{
\item userID integer user ID
}
Otherwise, send a JSON-formatted reply with the single item "error": "authorization failed"
and call stop(), which ends processing of the current request.
}
\description{
validate a request by looking up its token, failing gracefully
}

